refactor(downloader): retain owned runtime interfaces
