+++ rust-srec/docs/en/concepts/architecture.md
@@ -397,3 +397,9 @@ Exposed in `/api/health` as a single aggregated `output-root` component listing 
   - That marker keeps the oldest and newest unresolved generations plus a count of the ones in
     between, so a restart loop cannot grow it. Startup and exit messages report how many
     generations still owe recovery.
+
+## Downloader Rust Interfaces
+
+The download manager keeps event contracts in `downloader::manager::events`, acknowledged delivery in `coordination`, engine configuration in `configuration`, and attempt ownership in `attempt`. These implementation modules remain private; public event imports through `downloader` and internal imports through `downloader::manager` are preserved. Runtime shutdown uses `DownloadManager::shutdown_until`, and the active entry retains its queue slot until removal.
+
+The unused pending-download configuration update API and write-only retry override have been removed, along with `stop_all`, `get_downloads_by_status`, `set_high_priority_extra_slots` and the old process-waiter helpers. Rust integrations should use the supported manager shutdown and snapshot methods, `DownloadConfig::build_pipeline_config` / `build_hls_pipeline_config` / `build_flv_pipeline_config`, and the public `CircuitBreaker` methods. Internal breaker ownership uses `CircuitBreakerManager::get`. Mesio engine diagnostics report the linked crate's `mesio::VERSION`.
