+++ rust-srec/docs/zh/release-notes/unreleased.md
@@ -2,6 +2,12 @@
 
 ## `unreleased`
 
+## 下载器接口
+
+- **Mesio 诊断报告所链接库的版本**
+
+  引擎检查现在报告 Mesio 编译时的包版本，不再使用硬编码的历史版本。下载管理器的任务所有权、事件传递和测试已按职责拆分，移除了未使用的更新、进程及配置辅助接口。现有下载事件和运行时关闭行为保持不变；请参阅 [Rust 接口说明](../concepts/architecture.md#下载器-rust-接口)。
+
 ## 内部指标
 
 - **监控代码与实际可用接口保持一致**
