+++ rust-srec/src/downloader/engine/utils.rs
@@ -11,7 +11,6 @@ mod disk_full;
 mod ffmpeg_parser;
 mod files;
 mod output_record_reader;
-mod process_runner;
 mod redact;
 mod version_probe;
 
@@ -25,7 +24,6 @@ pub use ffmpeg_parser::{
 };
 pub use files::ensure_output_dir;
 pub use output_record_reader::OutputRecordReader;
-pub use process_runner::{spawn_piped_process_waiter, spawn_process_waiter};
 pub use redact::redact_process_args;
 pub(crate) use version_probe::{probe_version, probe_version_sync};
