+++ rust-srec/src/downloader/manager.rs
@@ -2,6 +2,20 @@
 
 mod attempt;
 mod configuration;
+mod coordination;
+mod events;
+
+#[cfg(test)]
+mod tests;
+
+use coordination::DownloadEventPublisher;
+pub(crate) use coordination::{
+    DownloadCoordinationReceiver, DownloadCoordinationSender, download_coordination_channel,
+};
+pub use events::{
+    ConfigUpdateType, DownloadManagerEvent, DownloadProgressEvent, DownloadRejectedKind,
+    DownloadShutdownReport, DownloadStopCause, DownloadTerminalEvent, EngineEndSignal,
+};
 
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -13,14 +27,13 @@ use dashmap::DashMap;
 use parking_lot::{Mutex, RwLock};
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
-use tokio::sync::{broadcast, mpsc, oneshot};
+use tokio::sync::broadcast;
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, info, warn};
 
 use super::engine::{
     DownloadConfig, DownloadEngine, DownloadFailureKind, DownloadHandle, DownloadInfo,
-    DownloadProgress, DownloadProtocol, DownloadStatus, EngineType, FfmpegEngine, IoErrorKindSer,
-    MesioEngine, StreamlinkEngine,
+    DownloadProgress, DownloadStatus, EngineType, FfmpegEngine, MesioEngine, StreamlinkEngine,
 };
 use super::output_root_gate::OutputRootGate;
 use super::queue::{
@@ -71,57 +84,6 @@ fn io_error_in_chain<'a>(err: &'a (dyn std::error::Error + 'static)) -> Option<&
         current = e.source();
     }
     None
-}
-
-/// Pending configuration update for an active download.
-///
-/// Stores configuration changes that will be applied when the next segment starts.
-/// Multiple updates can be merged, with newer values overwriting older ones.
-#[derive(Debug, Clone, Default)]
-pub struct PendingConfigUpdate {
-    /// Updated cookies (if any).
-    pub cookies: Option<String>,
-    /// Updated headers (if any).
-    pub headers: Option<Vec<(String, String)>>,
-    /// Updated retry configuration (if any).
-    pub retry_config: Option<RetryConfig>,
-    /// Timestamp when the update was queued.
-    pub queued_at: DateTime<Utc>,
-}
-
-impl PendingConfigUpdate {
-    /// Create a new pending config update with the current timestamp.
-    pub fn new(
-        cookies: Option<String>,
-        headers: Option<Vec<(String, String)>>,
-        retry_config: Option<RetryConfig>,
-    ) -> Self {
-        Self {
-            cookies,
-            headers,
-            retry_config,
-            queued_at: Utc::now(),
-        }
-    }
-
-    /// Check if there are any pending updates.
-    pub fn has_updates(&self) -> bool {
-        self.cookies.is_some() || self.headers.is_some() || self.retry_config.is_some()
-    }
-
-    /// Merge another update into this one (newer values overwrite).
-    pub fn merge(&mut self, other: PendingConfigUpdate) {
-        if other.cookies.is_some() {
-            self.cookies = other.cookies;
-        }
-        if other.headers.is_some() {
-            self.headers = other.headers;
-        }
-        if other.retry_config.is_some() {
-            self.retry_config = other.retry_config;
-        }
-        self.queued_at = other.queued_at;
-    }
 }
 
 /// Configuration for the Download Manager.
@@ -178,11 +140,9 @@ struct ActiveDownload {
     /// active map.
     #[expect(
         dead_code,
-        reason = "retained for optional runtime paths and diagnostics"
+        reason = "owns queue capacity until the active download is removed"
     )]
     slot: Option<ActiveSlot>,
-    /// Retry configuration override applied via config update.
-    retry_config_override: Option<RetryConfig>,
 }
 
 impl From<&ActiveDownload> for DownloadInfo {
@@ -221,216 +181,6 @@ impl AttemptPhase {
     }
 }
 
-#[derive(Clone)]
-struct DownloadEventPublisher {
-    observer_tx: broadcast::Sender<DownloadManagerEvent>,
-    coordination_tx: Option<DownloadCoordinationSender>,
-}
-
-impl DownloadEventPublisher {
-    fn new(
-        observer_tx: broadcast::Sender<DownloadManagerEvent>,
-        coordination_tx: Option<DownloadCoordinationSender>,
-    ) -> Self {
-        Self {
-            observer_tx,
-            coordination_tx,
-        }
-    }
-
-    /// Broadcast an event that no required consumer has to apply.
-    ///
-    /// Coordination-requiring events go through [`Self::publish_and_wait`] so
-    /// their acknowledgement is observed; routing one here would enqueue it and
-    /// then discard whether the consumer applied it, which is why this asserts
-    /// on the distinction rather than silently branching on it.
-    fn publish(&self, event: DownloadManagerEvent) -> bool {
-        debug_assert!(
-            !event.requires_coordination(),
-            "coordination-requiring events must use publish_and_wait"
-        );
-        self.observe(event)
-    }
-
-    async fn publish_and_wait(
-        &self,
-        event: DownloadManagerEvent,
-    ) -> std::result::Result<(), String> {
-        self.coordinate_and_wait(&event).await?;
-        self.observe(event);
-        Ok(())
-    }
-
-    async fn coordinate_and_wait(
-        &self,
-        event: &DownloadManagerEvent,
-    ) -> std::result::Result<(), String> {
-        if !event.requires_coordination() {
-            return Ok(());
-        }
-        match &self.coordination_tx {
-            Some(sender) => sender.publish(event.clone()).wait().await,
-            None => Ok(()),
-        }
-    }
-
-    fn observe(&self, event: DownloadManagerEvent) -> bool {
-        self.observer_tx.send(event).is_ok()
-    }
-
-    fn subscribe(&self) -> broadcast::Receiver<DownloadManagerEvent> {
-        self.observer_tx.subscribe()
-    }
-
-    async fn shutdown_coordination(&self) -> std::result::Result<(), &'static str> {
-        match &self.coordination_tx {
-            Some(sender) => sender.shutdown().await,
-            None => Ok(()),
-        }
-    }
-}
-
-enum DownloadCoordinationDelivery {
-    Event {
-        event: Box<DownloadManagerEvent>,
-        acknowledgement: oneshot::Sender<std::result::Result<(), String>>,
-    },
-    Shutdown(oneshot::Sender<()>),
-}
-
-#[derive(Clone)]
-pub(crate) struct DownloadCoordinationSender {
-    state: Arc<Mutex<DownloadCoordinationState>>,
-}
-
-struct DownloadCoordinationState {
-    tx: mpsc::UnboundedSender<DownloadCoordinationDelivery>,
-    accepting: bool,
-}
-
-impl DownloadCoordinationSender {
-    fn publish(&self, event: DownloadManagerEvent) -> DownloadCoordinationReceipt {
-        let (acknowledgement, acknowledged) = oneshot::channel();
-        let delivery = DownloadCoordinationDelivery::Event {
-            event: Box::new(event),
-            acknowledgement,
-        };
-        let mut state = self.state.lock();
-        if !state.accepting || state.tx.send(delivery).is_err() {
-            state.accepting = false;
-            DownloadCoordinationReceipt::Unavailable
-        } else {
-            DownloadCoordinationReceipt::Pending(acknowledged)
-        }
-    }
-
-    async fn shutdown(&self) -> std::result::Result<(), &'static str> {
-        let (acknowledgement, acknowledged) = oneshot::channel();
-        {
-            let mut state = self.state.lock();
-            if !state.accepting {
-                return Err("required download coordination channel is already closed");
-            }
-            state.accepting = false;
-            state
-                .tx
-                .send(DownloadCoordinationDelivery::Shutdown(acknowledgement))
-                .map_err(|_| "required download coordination channel is closed")?;
-        }
-        acknowledged
-            .await
-            .map_err(|_| "download coordination handler stopped before acknowledging shutdown")
-    }
-
-    #[cfg(test)]
-    pub(crate) async fn publish_and_wait_for_test(
-        &self,
-        event: DownloadManagerEvent,
-    ) -> std::result::Result<(), String> {
-        self.publish(event).wait().await
-    }
-
-    #[cfg(test)]
-    pub(crate) async fn shutdown_for_test(&self) -> std::result::Result<(), &'static str> {
-        self.shutdown().await
-    }
-}
-
-enum DownloadCoordinationReceipt {
-    Pending(oneshot::Receiver<std::result::Result<(), String>>),
-    Unavailable,
-}
-
-impl DownloadCoordinationReceipt {
-    async fn wait(self) -> std::result::Result<(), String> {
-        match self {
-            Self::Pending(acknowledged) => acknowledged.await.map_err(|_| {
-                "download coordination handler stopped before acknowledging event".to_string()
-            })?,
-            Self::Unavailable => {
-                Err("required download coordination channel is closed".to_string())
-            }
-        }
-    }
-}
-
-pub(crate) struct DownloadCoordinationEvent {
-    event: DownloadManagerEvent,
-    acknowledgement: oneshot::Sender<std::result::Result<(), String>>,
-}
-
-impl DownloadCoordinationEvent {
-    pub(crate) fn into_parts(
-        self,
-    ) -> (
-        DownloadManagerEvent,
-        oneshot::Sender<std::result::Result<(), String>>,
-    ) {
-        (self.event, self.acknowledgement)
-    }
-}
-
-pub(crate) struct DownloadCoordinationReceiver {
-    rx: mpsc::UnboundedReceiver<DownloadCoordinationDelivery>,
-}
-
-impl DownloadCoordinationReceiver {
-    pub(crate) async fn recv(
-        &mut self,
-    ) -> std::result::Result<Option<DownloadCoordinationEvent>, &'static str> {
-        match self.rx.recv().await {
-            Some(DownloadCoordinationDelivery::Event {
-                event,
-                acknowledgement,
-            }) => Ok(Some(DownloadCoordinationEvent {
-                event: *event,
-                acknowledgement,
-            })),
-            Some(DownloadCoordinationDelivery::Shutdown(acknowledgement)) => {
-                if acknowledgement.send(()).is_err() {
-                    debug!("Download coordination shutdown acknowledgement receiver was dropped");
-                }
-                Ok(None)
-            }
-            None => Err("required download coordination channel closed without shutdown"),
-        }
-    }
-}
-
-pub(crate) fn download_coordination_channel()
--> (DownloadCoordinationSender, DownloadCoordinationReceiver) {
-    let (tx, rx) = mpsc::unbounded_channel();
-    (
-        DownloadCoordinationSender {
-            state: Arc::new(Mutex::new(DownloadCoordinationState {
-                tx,
-                accepting: true,
-            })),
-        },
-        DownloadCoordinationReceiver { rx },
-    )
-}
-
 /// The Download Manager service.
 pub struct DownloadManager {
     /// Configuration.
@@ -447,8 +197,6 @@ pub struct DownloadManager {
     /// against the required-event shutdown marker.
     operation_gate: tokio::sync::RwLock<()>,
     accepting_operations: AtomicBool,
-    /// Pending configuration updates keyed by download_id.
-    pending_updates: Arc<DashMap<String, PendingConfigUpdate>>,
     /// Next session-scoped segment index keyed by recording session id.
     /// The per-download `engine_segment_index -> session_segment_index`
     /// mapping is held as a local variable in the spawn loop in
@@ -480,485 +228,6 @@ pub struct DownloadManager {
     /// hot path, hence `AtomicI64` rather than the `RwLock`-guarded
     /// [`DownloadManagerConfig`].
     queue_freshness_threshold_ms: AtomicI64,
-}
-
-/// Type of configuration that was updated.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ConfigUpdateType {
-    /// Only cookies were updated.
-    Cookies,
-    /// Only headers were updated.
-    Headers,
-    /// Only retry configuration was updated.
-    RetryConfig,
-    /// Multiple configuration types were updated.
-    Multiple,
-}
-
-/// How the engine observed the end of a stream when emitting
-/// [`DownloadTerminalEvent::Completed`].
-///
-/// Different signals carry different confidence about whether the upstream
-/// stream is *actually* over vs. whether we just got disconnected and the
-/// streamer will reappear with a fresh URL.
-///
-/// Consumers (today: [`crate::session::SessionLifecycle`]) use this to choose
-/// between firing the session-complete pipeline immediately and entering a
-/// hysteresis quiet-period that absorbs reconnects.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub enum EngineEndSignal {
-    /// HLS playlist contained `#EXT-X-ENDLIST`. The platform itself marked
-    /// the stream as complete — definitively over.
-    HlsEndlist,
-    /// Connection closed cleanly with no explicit end marker. Could be EOF,
-    /// could be a reconnect-friendly drop. Ambiguous; the lifecycle should
-    /// hold the session in hysteresis to absorb a possible resume.
-    ///
-    /// Used by mesio FLV (TCP close), mesio HLS without `#EXT-X-ENDLIST`,
-    /// and any other engine that observes a clean disconnect without a
-    /// platform-asserted end marker.
-    CleanDisconnect,
-    /// Subprocess (ffmpeg / streamlink) exited with status 0. Ambiguous —
-    /// could be EOF, could be the process being killed cleanly externally.
-    SubprocessExitZero,
-    /// Engine doesn't expose a finer signal. Treat as non-authoritative
-    /// (default for back-fill / unknown engines).
-    #[default]
-    Unknown,
-}
-
-impl EngineEndSignal {
-    /// Whether this signal alone is sufficient to mark the stream as done
-    /// without waiting for hysteresis. Today, only HLS `#EXT-X-ENDLIST` is.
-    pub fn is_authoritative(&self) -> bool {
-        matches!(self, Self::HlsEndlist)
-    }
-
-    /// Short, stable label for logging / metrics.
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::HlsEndlist => "hls_endlist",
-            Self::CleanDisconnect => "clean_disconnect",
-            Self::SubprocessExitZero => "subprocess_exit_zero",
-            Self::Unknown => "unknown",
-        }
-    }
-}
-
-/// Reason why a download was stopped.
-///
-/// Used to disambiguate user cancellation from internal orchestration stops.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DownloadStopCause {
-    /// User explicitly cancelled the download (typically implies "stop monitoring").
-    User,
-    /// Stream was determined to be offline (end-of-stream).
-    StreamerOffline,
-    /// Danmu stream emitted `StreamClosed` and we stop the download promptly.
-    DanmuStreamClosed,
-    /// The streamer is live but recording is no longer allowed (schedule window ended).
-    OutOfSchedule,
-    /// Streamer was disabled/deleted; downloads are stopped as part of cleanup.
-    StreamerDisabled,
-    /// Application shutdown.
-    Shutdown,
-    /// Other internal/system stop reason.
-    Other(String),
-}
-
-/// Result of a graceful download-runtime shutdown.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DownloadShutdownReport {
-    /// Downloads that were active when shutdown fenced new attempts.
-    pub stopped_download_ids: Vec<String>,
-    /// Attempts that exceeded the graceful deadline but were still joined
-    /// before shutdown continued.
-    pub deadline_exceeded_download_ids: Vec<String>,
-    /// Whether the required coordination consumer acknowledged its final
-    /// marker. Producers are quiesced even when this is false.
-    pub coordination_drained: bool,
-    /// Lifecycle errors from attempts that ended earlier in the run. Reported
-    /// and then ignored: they describe those recordings, not this shutdown.
-    pub runtime_failures: Vec<String>,
-    /// Attempt or coordination failures of the shutdown phase itself. Each one
-    /// means shutdown could not prove an attempt finalized, so callers treat
-    /// them as fatal.
-    pub failures: Vec<String>,
-    /// Soft-budget overruns that were still contained to completion. Reported
-    /// so operators can raise the budget, never fatal.
-    pub overruns: Vec<String>,
-}
-
-impl DownloadStopCause {
-    pub fn as_str(&self) -> &str {
-        match self {
-            Self::User => "user",
-            Self::StreamerOffline => "streamer_offline",
-            Self::DanmuStreamClosed => "danmu_stream_closed",
-            Self::OutOfSchedule => "out_of_schedule",
-            Self::StreamerDisabled => "streamer_disabled",
-            Self::Shutdown => "shutdown",
-            Self::Other(_) => "other",
-        }
-    }
-}
-
-/// Events emitted by the Download Manager.
-///
-/// Events are grouped into two categories:
-///
-/// - [`DownloadProgressEvent`]: in-flight notifications about an ongoing
-///   download (start, progress, segment lifecycle, config changes). These do
-///   not mark a session as finished.
-/// - [`DownloadTerminalEvent`]: the download has stopped and no further events
-///   for the same `(download_id, session_id)` pair will be emitted. Consumers
-///   that need to react to "session ended" should match on this variant and
-///   use [`DownloadTerminalEvent::should_run_session_complete_pipeline`] to
-///   decide whether to fire the post-recording pipeline.
-///
-/// The split exists to make session-termination a first-class, non-droppable
-/// signal: Rust's exhaustive pattern matching forces every consumer to make
-/// an explicit decision for every terminal variant, where a flat enum would
-/// let a `_ => {}` catch-all silently drop one.
-#[derive(Debug, Clone)]
-pub enum DownloadManagerEvent {
-    Progress(DownloadProgressEvent),
-    Terminal(DownloadTerminalEvent),
-}
-
-/// Non-terminal download notifications.
-#[derive(Debug, Clone)]
-pub enum DownloadProgressEvent {
-    /// Download is parked waiting for a concurrency slot. Emitted only
-    /// when the request had to wait (fast-path acquires emit
-    /// [`Self::DownloadStarted`] directly with no preceding `DownloadQueued`).
-    /// Cleared by a subsequent `DownloadStarted` for the same session
-    /// when the engine starts, or by [`Self::DownloadDequeued`] if
-    /// the pipeline aborts before then.
-    DownloadQueued {
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        engine_type: EngineType,
-        is_high_priority: bool,
-        queued_at_ms: i64,
-    },
-    /// Cleanup signal emitted when a previously-queued pipeline
-    /// aborts before starting a download (cancellation, shutdown,
-    /// freshness/state re-check, etc.).
-    /// Not emitted for the fast-path acquire (which never emitted
-    /// `DownloadQueued`). Frontend clears the streamer's "queued"
-    /// badge on receipt; no scheduler action is taken — the
-    /// streamer's actual state machine is driven elsewhere.
-    DownloadDequeued {
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-    },
-    /// Download started.
-    DownloadStarted {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        engine_type: EngineType,
-        cdn_host: String,
-        download_url: String,
-    },
-    /// Progress update for a download.
-    Progress {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        status: DownloadStatus,
-        progress: DownloadProgress,
-    },
-    /// Segment started - a new segment file has been opened for writing.
-    SegmentStarted {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        segment_path: String,
-        segment_index: u32,
-        started_at: DateTime<Utc>,
-    },
-    /// Segment completed.
-    SegmentCompleted {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        segment_path: String,
-        segment_index: u32,
-        started_at: Option<DateTime<Utc>>,
-        completed_at: DateTime<Utc>,
-        duration_secs: f64,
-        size_bytes: u64,
-        split_reason_code: Option<String>,
-        split_reason_details_json: Option<String>,
-    },
-    /// Configuration was updated for a download.
-    ConfigUpdated {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        update_type: ConfigUpdateType,
-    },
-    /// Configuration update failed to apply.
-    ConfigUpdateFailed {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        error: String,
-    },
-}
-
-/// Terminal download notifications: the download has stopped, no further
-/// events for the same `(download_id, session_id)` pair will be emitted.
-///
-/// Consumers that need to react to "session ended" (notifications, pipeline
-/// scheduling, DB status updates, …) should centralise on this enum and call
-/// [`Self::should_run_session_complete_pipeline`] rather than re-deriving
-/// the policy per site.
-#[derive(Debug, Clone)]
-pub enum DownloadTerminalEvent {
-    /// Download completed normally — all segments flushed, outputs finalised.
-    ///
-    /// Important: a `Completed` event is *not* by itself authoritative
-    /// proof that the upstream stream is over. A clean TCP close or a
-    /// subprocess exiting with status 0 can mean either "EOF" or "we
-    /// reconnected." Engines pass an [`EngineEndSignal`] hint with the
-    /// event so [`crate::session::SessionLifecycle`] can tell HLS
-    /// `#EXT-X-ENDLIST` (definitively over) apart from FLV clean
-    /// disconnect (might be a reconnect-friendly drop).
-    Completed {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        total_bytes: u64,
-        total_duration_secs: f64,
-        total_segments: u32,
-        file_path: Option<String>,
-        /// How the engine observed the end. Lifecycle reads this to decide
-        /// whether to enter hysteresis (clean disconnect / subprocess exit
-        /// → ambiguous) or commit Ended directly (HLS endlist → authoritative).
-        engine_signal: EngineEndSignal,
-        /// External stop request that preceded the engine's clean completion.
-        ///
-        /// Completion describes the finalized recording output; this cause
-        /// preserves the control-plane intent for scheduler state handling.
-        stop_cause: Option<DownloadStopCause>,
-    },
-    /// Download failed — the engine gave up. Whatever output is on disk is
-    /// final; no more segments will arrive.
-    Failed {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        engine_type: EngineType,
-        /// Selected stream protocol. Mesio uses one engine type for both
-        /// HLS and FLV, so lifecycle needs this to classify failures without
-        /// guessing.
-        protocol: DownloadProtocol,
-        kind: DownloadFailureKind,
-        error: String,
-        recoverable: bool,
-    },
-    /// Download cancelled after the engine flushed its in-flight segment and
-    /// all preceding segment events were translated. Cancellation preserves
-    /// the logical session so a later attempt can resume it.
-    Cancelled {
-        download_id: String,
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        cause: DownloadStopCause,
-    },
-    /// Download was rejected before starting (e.g., circuit breaker open,
-    /// output-root filesystem unwritable).
-    ///
-    /// Unlike [`Self::Failed`], this indicates the download never started.
-    /// No `download_id` is available because the download was never created.
-    Rejected {
-        streamer_id: String,
-        streamer_name: String,
-        session_id: String,
-        reason: String,
-        /// How long to wait before retrying (cooldown of whichever subsystem
-        /// rejected the download).
-        retry_after_secs: Option<u64>,
-        /// Why the download was rejected. Carries the payload the scheduler
-        /// needs to route this to the correct
-        /// [`DownloadEndPolicy`](crate::scheduler::actor::DownloadEndPolicy) variant
-        /// and, ultimately, the correct [`crate::monitor::InfraBlockReason`].
-        kind: DownloadRejectedKind,
-    },
-}
-
-impl DownloadManagerEvent {
-    pub(crate) fn requires_coordination(&self) -> bool {
-        matches!(
-            self,
-            Self::Progress(
-                DownloadProgressEvent::SegmentStarted { .. }
-                    | DownloadProgressEvent::SegmentCompleted { .. }
-            ) | Self::Terminal(_)
-        )
-    }
-
-    /// Streamer id shared across both progress and terminal event shapes.
-    pub fn streamer_id(&self) -> &str {
-        match self {
-            Self::Progress(p) => p.streamer_id(),
-            Self::Terminal(t) => t.streamer_id(),
-        }
-    }
-
-    /// Streamer display name shared across both shapes.
-    pub fn streamer_name(&self) -> &str {
-        match self {
-            Self::Progress(p) => p.streamer_name(),
-            Self::Terminal(t) => t.streamer_name(),
-        }
-    }
-
-    /// Recording session id. Present on every variant.
-    pub fn session_id(&self) -> &str {
-        match self {
-            Self::Progress(p) => p.session_id(),
-            Self::Terminal(t) => t.session_id(),
-        }
-    }
-}
-
-impl DownloadProgressEvent {
-    pub fn streamer_id(&self) -> &str {
-        match self {
-            Self::DownloadQueued { streamer_id, .. }
-            | Self::DownloadDequeued { streamer_id, .. }
-            | Self::DownloadStarted { streamer_id, .. }
-            | Self::Progress { streamer_id, .. }
-            | Self::SegmentStarted { streamer_id, .. }
-            | Self::SegmentCompleted { streamer_id, .. }
-            | Self::ConfigUpdated { streamer_id, .. }
-            | Self::ConfigUpdateFailed { streamer_id, .. } => streamer_id,
-        }
-    }
-
-    pub fn streamer_name(&self) -> &str {
-        match self {
-            Self::DownloadQueued { streamer_name, .. }
-            | Self::DownloadDequeued { streamer_name, .. }
-            | Self::DownloadStarted { streamer_name, .. }
-            | Self::Progress { streamer_name, .. }
-            | Self::SegmentStarted { streamer_name, .. }
-            | Self::SegmentCompleted { streamer_name, .. }
-            | Self::ConfigUpdated { streamer_name, .. }
-            | Self::ConfigUpdateFailed { streamer_name, .. } => streamer_name,
-        }
-    }
-
-    /// `session_id` is always present on live download events, but `ConfigUpdated`
-    /// and `ConfigUpdateFailed` carry none today (they're scoped to a
-    /// `download_id`). Returns an empty string for those variants — callers
-    /// that need a real session id should only call this on variants that
-    /// have one.
-    pub fn session_id(&self) -> &str {
-        match self {
-            Self::DownloadQueued { session_id, .. }
-            | Self::DownloadDequeued { session_id, .. }
-            | Self::DownloadStarted { session_id, .. }
-            | Self::Progress { session_id, .. }
-            | Self::SegmentStarted { session_id, .. }
-            | Self::SegmentCompleted { session_id, .. } => session_id,
-            Self::ConfigUpdated { .. } | Self::ConfigUpdateFailed { .. } => "",
-        }
-    }
-}
-
-impl DownloadTerminalEvent {
-    pub fn streamer_id(&self) -> &str {
-        match self {
-            Self::Completed { streamer_id, .. }
-            | Self::Failed { streamer_id, .. }
-            | Self::Cancelled { streamer_id, .. }
-            | Self::Rejected { streamer_id, .. } => streamer_id,
-        }
-    }
-
-    pub fn streamer_name(&self) -> &str {
-        match self {
-            Self::Completed { streamer_name, .. }
-            | Self::Failed { streamer_name, .. }
-            | Self::Cancelled { streamer_name, .. }
-            | Self::Rejected { streamer_name, .. } => streamer_name,
-        }
-    }
-
-    pub fn session_id(&self) -> &str {
-        match self {
-            Self::Completed { session_id, .. }
-            | Self::Failed { session_id, .. }
-            | Self::Cancelled { session_id, .. }
-            | Self::Rejected { session_id, .. } => session_id,
-        }
-    }
-
-    /// `download_id` is present for every terminal variant except
-    /// [`Self::Rejected`] (rejection happens before the download is created).
-    pub fn download_id(&self) -> Option<&str> {
-        match self {
-            Self::Completed { download_id, .. }
-            | Self::Failed { download_id, .. }
-            | Self::Cancelled { download_id, .. } => Some(download_id),
-            Self::Rejected { .. } => None,
-        }
-    }
-
-    /// Whether this termination represents the recording session reaching a
-    /// final state with outputs on disk that are ready for post-processing.
-    ///
-    /// - [`Self::Completed`]: `true` — normal end, outputs finalised.
-    /// - [`Self::Failed`]: `true` — the engine gave up; whatever's on disk
-    ///   is final. Prior to this method existing, sessions that ended this
-    ///   way silently skipped the session-complete pipeline.
-    /// - [`Self::Cancelled`]: `false` — the engine did not report a clean
-    ///   completion. The control-plane stop owner decides whether to end or
-    ///   preserve the session (shutdown preserves it for restart).
-    /// - [`Self::Rejected`]: `false` — the download never started, no
-    ///   outputs exist.
-    pub fn should_run_session_complete_pipeline(&self) -> bool {
-        matches!(self, Self::Completed { .. } | Self::Failed { .. })
-    }
-}
-
-/// Reason a [`DownloadTerminalEvent::Rejected`] event was emitted.
-///
-/// Distinct from the free-form `reason` string because the scheduler needs
-/// structured data to decide which state to transition the streamer into —
-/// circuit-breaker blocks go to `TemporalDisabled`, gate blocks go to
-/// `OutOfSpace`.
-#[derive(Debug, Clone)]
-pub enum DownloadRejectedKind {
-    /// Engine circuit breaker is open for the resolved engine.
-    CircuitBreaker,
-    /// Output-root write gate has the target filesystem in the Degraded state.
-    /// The payload mirrors the gate's `GateBlocked` so the scheduler can
-    /// construct an [`crate::monitor::InfraBlockReason::OutputRootUnavailable`].
-    OutputRootUnavailable {
-        path: std::path::PathBuf,
-        io_kind: IoErrorKindSer,
-    },
-    /// The streamer is inside its error-backoff window (`disabled_until` set
-    /// by `MonitorService::handle_error`). Emitted by the container's
-    /// download-start gate rather than `preflight` — the backoff itself is
-    /// already persisted, so the scheduler only reschedules the actor after
-    /// `retry_after_secs`; no `InfraBlockReason` write is needed.
-    StreamerBackoff,
 }
 
 /// Input to [`DownloadManager::preflight`].
@@ -1029,7 +298,6 @@ impl DownloadManager {
             attempts: AttemptSupervisor::new(),
             operation_gate: tokio::sync::RwLock::new(()),
             accepting_operations: AtomicBool::new(true),
-            pending_updates: Arc::new(DashMap::new()),
             session_segment_indices: Arc::new(DashMap::new()),
             engines: RwLock::new(HashMap::new()),
             circuit_breakers,
@@ -1607,11 +875,8 @@ impl DownloadManager {
     /// `AcquireError::ShuttingDown`; pending acquires are notified and
     /// return the same error rather than waiting for a slot.
     ///
-    /// Call this BEFORE [`Self::stop_all`] during graceful shutdown so
-    /// queued pipelines released by `stop_all`'s slot drops don't try
-    /// to spin up new engines as the rest of the system tears down.
-    /// Active downloads aren't affected — they continue until
-    /// `stop_all` drops them.
+    /// Runtime shutdown uses [`Self::shutdown_until`] to fence admission and
+    /// drain the active attempts after closing the queue.
     pub fn shutdown_queue(&self) {
         self.queue.shutdown();
     }
@@ -1648,16 +913,6 @@ impl DownloadManager {
         }
 
         self.queue.set_normal_capacity(limit)
-    }
-
-    /// Adjust the number of high-priority extra slots at runtime (0 disables high-priority slots).
-    pub fn set_high_priority_extra_slots(&self, slots: usize) -> usize {
-        {
-            let mut config = self.config.write();
-            config.high_priority_extra_slots = slots;
-        }
-
-        self.queue.set_high_extra_capacity(slots)
     }
 
     /// Current queue-wait freshness threshold in milliseconds.
@@ -1762,82 +1017,6 @@ impl DownloadManager {
         Ok(guard)
     }
 
-    /// Update configuration for an active download.
-    ///
-    /// Queues configuration updates (cookies, headers, retry policy) to be applied
-    /// when the next segment starts. Multiple updates are merged, with newer values
-    /// overwriting older ones.
-    ///
-    /// # Arguments
-    /// * `download_id` - The ID of the download to update
-    /// * `cookies` - Optional new cookies to apply
-    /// * `headers` - Optional new headers to apply
-    /// * `retry_config` - Optional new retry configuration to apply
-    ///
-    /// # Returns
-    /// * `Ok(())` if the update was queued successfully
-    /// * `Err(NotFound)` if the download does not exist
-    pub fn update_download_config(
-        &self,
-        download_id: &str,
-        cookies: Option<String>,
-        headers: Option<Vec<(String, String)>>,
-        retry_config: Option<RetryConfig>,
-    ) -> Result<()> {
-        // Validate download exists in active_downloads
-        let download =
-            self.active_downloads
-                .get(download_id)
-                .ok_or_else(|| crate::Error::NotFound {
-                    entity_type: "Download".to_string(),
-                    id: download_id.to_string(),
-                })?;
-
-        let streamer_id = download.handle.config.read().streamer_id.clone();
-        // Drop the reference to avoid holding the lock while updating pending_updates
-        drop(download);
-
-        // Capture presence flags for logging before the payloads move into
-        // the pending update.
-        let has_cookies = cookies.is_some();
-        let has_headers = headers.is_some();
-        let has_retry = retry_config.is_some();
-
-        // Create the new pending update
-        let new_update = PendingConfigUpdate::new(cookies, headers, retry_config);
-
-        // Only store if there are actual updates
-        if new_update.has_updates() {
-            // Create or merge PendingConfigUpdate in pending_updates map
-            match self.pending_updates.entry(download_id.to_string()) {
-                dashmap::mapref::entry::Entry::Occupied(mut existing) => {
-                    existing.get_mut().merge(new_update);
-                }
-                dashmap::mapref::entry::Entry::Vacant(slot) => {
-                    slot.insert(new_update);
-                }
-            }
-
-            // Log the queued update
-            info!(
-                "Config update queued for download {}: cookies={}, headers={}, retry={}",
-                download_id, has_cookies, has_headers, has_retry
-            );
-
-            debug!(
-                "Download {} for streamer {} will apply config on next segment",
-                download_id, streamer_id
-            );
-        } else {
-            debug!(
-                "Empty config update for download {} - no changes queued",
-                download_id
-            );
-        }
-
-        Ok(())
-    }
-
     /// Get download by streamer ID.
     pub fn get_download_by_streamer(&self, streamer_id: &str) -> Option<DownloadInfo> {
         self.active_downloads
@@ -1864,253 +1043,6 @@ impl DownloadManager {
                     DownloadStatus::Starting | DownloadStatus::Downloading
                 )
         })
-    }
-
-    /// Take pending updates for a download (called by engines at segment boundaries).
-    ///
-    /// Atomically removes and returns the pending configuration update for the specified
-    /// download. This should be called by download engines when starting a new segment
-    /// to apply any queued configuration changes.
-    ///
-    /// # Arguments
-    /// * `download_id` - The ID of the download to take pending updates for
-    ///
-    /// # Returns
-    /// * `Some(PendingConfigUpdate)` if there were pending updates
-    /// * `None` if no updates were pending for this download
-    pub fn take_pending_updates(&self, download_id: &str) -> Option<PendingConfigUpdate> {
-        self.pending_updates
-            .remove(download_id)
-            .map(|(_, update)| update)
-    }
-
-    /// Check if a download has pending configuration updates.
-    ///
-    /// # Arguments
-    /// * `download_id` - The ID of the download to check
-    ///
-    /// # Returns
-    /// * `true` if there are pending updates for this download
-    /// * `false` otherwise
-    pub fn has_pending_updates(&self, download_id: &str) -> bool {
-        self.pending_updates.contains_key(download_id)
-    }
-
-    /// Emit a ConfigUpdated event for a successfully applied configuration update.
-    ///
-    /// This helper method determines the appropriate `ConfigUpdateType` based on which
-    /// fields were present in the `PendingConfigUpdate` and emits the event via the
-    /// broadcast channel.
-    ///
-    /// # Arguments
-    /// * `download_id` - The ID of the download that was updated
-    /// * `streamer_id` - The streamer ID associated with the download
-    /// * `update` - The pending config update that was applied
-    ///
-    /// # Returns
-    /// * `true` if the event was sent successfully (at least one receiver)
-    /// * `false` if there were no receivers or the update had no changes
-    pub fn emit_config_updated(
-        &self,
-        download_id: &str,
-        streamer_id: &str,
-        update: &PendingConfigUpdate,
-    ) -> bool {
-        // Don't emit if there are no actual updates
-        if !update.has_updates() {
-            return false;
-        }
-
-        let update_type = Self::determine_config_update_type(update);
-
-        let streamer_name = self
-            .active_downloads
-            .get(download_id)
-            .map(|d| d.handle.config.read().streamer_name.clone())
-            .unwrap_or_else(|| streamer_id.to_string());
-
-        let event = DownloadManagerEvent::Progress(DownloadProgressEvent::ConfigUpdated {
-            download_id: download_id.to_string(),
-            streamer_id: streamer_id.to_string(),
-            streamer_name,
-            update_type,
-        });
-
-        // Broadcast send returns Ok if at least one receiver got the message
-        // Returns Err if there are no receivers, which is fine
-        if self.events.publish(event) {
-            debug!(
-                "Emitted ConfigUpdated event for download {} (streamer {})",
-                download_id, streamer_id
-            );
-            true
-        } else {
-            // No receivers - this is not an error, just means no one is listening
-            debug!(
-                "ConfigUpdated event for download {} had no receivers",
-                download_id
-            );
-            false
-        }
-    }
-
-    /// Emit a ConfigUpdateFailed event when a configuration update fails to apply.
-    ///
-    /// # Arguments
-    /// * `download_id` - The ID of the download that failed to update
-    /// * `streamer_id` - The streamer ID associated with the download
-    /// * `error` - Description of the error that occurred
-    ///
-    /// # Returns
-    /// * `true` if the event was sent successfully (at least one receiver)
-    /// * `false` if there were no receivers
-    pub fn emit_config_update_failed(
-        &self,
-        download_id: &str,
-        streamer_id: &str,
-        error: &str,
-    ) -> bool {
-        let streamer_name = self
-            .active_downloads
-            .get(download_id)
-            .map(|d| d.handle.config.read().streamer_name.clone())
-            .unwrap_or_else(|| streamer_id.to_string());
-
-        let event = DownloadManagerEvent::Progress(DownloadProgressEvent::ConfigUpdateFailed {
-            download_id: download_id.to_string(),
-            streamer_id: streamer_id.to_string(),
-            streamer_name,
-            error: error.to_string(),
-        });
-
-        if self.events.publish(event) {
-            warn!(
-                "Emitted ConfigUpdateFailed event for download {}: {}",
-                download_id, error
-            );
-            true
-        } else {
-            debug!(
-                "ConfigUpdateFailed event for download {} had no receivers",
-                download_id
-            );
-            false
-        }
-    }
-
-    /// Determine the ConfigUpdateType based on which fields are present in the update.
-    ///
-    /// # Arguments
-    /// * `update` - The pending config update to analyze
-    ///
-    /// # Returns
-    /// The appropriate `ConfigUpdateType` variant:
-    /// - `Multiple` if more than one field is set
-    /// - `Cookies`, `Headers`, or `RetryConfig` if only one field is set
-    /// - `Multiple` as fallback (should not happen if `has_updates()` is true)
-    fn determine_config_update_type(update: &PendingConfigUpdate) -> ConfigUpdateType {
-        let has_cookies = update.cookies.is_some();
-        let has_headers = update.headers.is_some();
-        let has_retry = update.retry_config.is_some();
-
-        let count = [has_cookies, has_headers, has_retry]
-            .iter()
-            .filter(|&&x| x)
-            .count();
-
-        if count > 1 {
-            ConfigUpdateType::Multiple
-        } else if has_cookies {
-            ConfigUpdateType::Cookies
-        } else if has_headers {
-            ConfigUpdateType::Headers
-        } else if has_retry {
-            ConfigUpdateType::RetryConfig
-        } else {
-            // Fallback - should not happen if has_updates() returned true
-            ConfigUpdateType::Multiple
-        }
-    }
-
-    fn apply_pending_update_to_download(
-        download: &mut ActiveDownload,
-        update: PendingConfigUpdate,
-        download_id: &str,
-        streamer_id: &str,
-        events: &DownloadEventPublisher,
-    ) {
-        // Classify before destructuring so the payloads can move into the
-        // handle config without a full `PendingConfigUpdate` clone.
-        let update_type = Self::determine_config_update_type(&update);
-        let mut applied = false;
-        let PendingConfigUpdate {
-            cookies,
-            headers,
-            retry_config,
-            ..
-        } = update;
-
-        if cookies.is_some() || headers.is_some() {
-            let mut cfg = download.handle.config.write();
-            if let Some(cookie_val) = cookies {
-                cfg.cookies = Some(cookie_val);
-                applied = true;
-            }
-            if let Some(header_val) = headers {
-                cfg.headers = header_val;
-                applied = true;
-            }
-        }
-
-        if let Some(retry) = retry_config {
-            download.retry_config_override = Some(retry);
-            applied = true;
-        }
-
-        if applied {
-            let streamer_name = download.handle.config.read().streamer_name.clone();
-            events.publish(DownloadManagerEvent::Progress(
-                DownloadProgressEvent::ConfigUpdated {
-                    download_id: download_id.to_string(),
-                    streamer_id: streamer_id.to_string(),
-                    streamer_name,
-                    update_type,
-                },
-            ));
-        }
-    }
-
-    /// Get downloads by status.
-    pub fn get_downloads_by_status(&self, status: DownloadStatus) -> Vec<DownloadInfo> {
-        self.active_downloads
-            .iter()
-            .filter(|entry| entry.value().status == status)
-            .map(|entry| DownloadInfo::from(entry.value()))
-            .collect()
-    }
-
-    /// Stop all active downloads.
-    pub async fn stop_all(&self) -> Vec<String> {
-        let download_ids: Vec<String> = self
-            .active_downloads
-            .iter()
-            .map(|entry| entry.key().clone())
-            .collect();
-
-        let results = futures::future::join_all(
-            download_ids
-                .iter()
-                .map(|id| self.stop_download_with_reason(id, DownloadStopCause::Shutdown)),
-        )
-        .await;
-        let stopped = download_ids
-            .into_iter()
-            .zip(results)
-            .filter_map(|(id, result)| result.is_ok().then_some(id))
-            .collect::<Vec<_>>();
-
-        info!("Stopped {} downloads", stopped.len());
-        stopped
     }
 
     /// Fence new attempts, request a graceful stop for every active download,
@@ -2212,1864 +1144,5 @@ impl DownloadManager {
 impl Default for DownloadManager {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::time::Duration;
-
-    use crate::downloader::SegmentInfo;
-    use crate::downloader::engine::{EngineStartError, SegmentEvent};
-
-    use super::*;
-    use async_trait::async_trait;
-    use std::sync::atomic::Ordering;
-
-    #[derive(Clone)]
-    struct ScriptedSegmentEngine {
-        prelude: Vec<SegmentEvent>,
-        /// Optional barrier: when set, the engine emits `prelude`, then awaits
-        /// this notify before emitting `tail` + `DownloadCompleted`. Lets a
-        /// test interleave `stop_download` between the prelude and the tail
-        /// to reproduce the trailing-flush race.
-        release_tail: Option<Arc<tokio::sync::Notify>>,
-        release_tail_on_cancel: bool,
-        tail: Vec<SegmentEvent>,
-    }
-
-    impl ScriptedSegmentEngine {
-        fn new(events: Vec<SegmentEvent>) -> Self {
-            Self {
-                prelude: events,
-                release_tail: None,
-                release_tail_on_cancel: false,
-                tail: Vec::new(),
-            }
-        }
-
-        fn with_gated_tail(
-            prelude: Vec<SegmentEvent>,
-            release: Arc<tokio::sync::Notify>,
-            tail: Vec<SegmentEvent>,
-        ) -> Self {
-            Self {
-                prelude,
-                release_tail: Some(release),
-                release_tail_on_cancel: false,
-                tail,
-            }
-        }
-
-        fn with_shutdown_tail(prelude: Vec<SegmentEvent>, tail: Vec<SegmentEvent>) -> Self {
-            Self {
-                prelude,
-                release_tail: None,
-                release_tail_on_cancel: true,
-                tail,
-            }
-        }
-    }
-
-    #[async_trait]
-    impl DownloadEngine for ScriptedSegmentEngine {
-        fn engine_type(&self) -> EngineType {
-            EngineType::Ffmpeg
-        }
-
-        async fn run(
-            &self,
-            handle: Arc<DownloadHandle>,
-        ) -> std::result::Result<(), EngineStartError> {
-            for event in self.prelude.clone() {
-                handle
-                    .event_tx
-                    .send(event)
-                    .await
-                    .map_err(|e| EngineStartError {
-                        kind: DownloadFailureKind::Other,
-                        message: format!("failed to emit scripted segment event: {}", e),
-                    })?;
-            }
-            if let Some(release) = self.release_tail.as_ref() {
-                release.notified().await;
-            } else if self.release_tail_on_cancel {
-                handle.cancellation_token.cancelled().await;
-            }
-            for event in self.tail.clone() {
-                handle
-                    .event_tx
-                    .send(event)
-                    .await
-                    .map_err(|e| EngineStartError {
-                        kind: DownloadFailureKind::Other,
-                        message: format!("failed to emit scripted tail event: {}", e),
-                    })?;
-            }
-            handle
-                .event_tx
-                .send(SegmentEvent::DownloadCompleted {
-                    total_bytes: 0,
-                    total_duration_secs: 0.0,
-                    total_segments: 0,
-                    engine_signal: EngineEndSignal::Unknown,
-                })
-                .await
-                .map_err(|e| EngineStartError {
-                    kind: DownloadFailureKind::Other,
-                    message: format!("failed to emit scripted completion event: {}", e),
-                })?;
-            Ok(())
-        }
-
-        fn is_available(&self) -> bool {
-            true
-        }
-
-        fn version(&self) -> Option<String> {
-            Some("scripted".to_string())
-        }
-    }
-
-    fn completed_segment(path: std::path::PathBuf, index: u32) -> SegmentEvent {
-        SegmentEvent::SegmentCompleted(SegmentInfo {
-            path,
-            duration_secs: 1.0,
-            size_bytes: 128,
-            index,
-            started_at: None,
-            completed_at: Utc::now(),
-            split_reason_code: None,
-            split_reason_details_json: None,
-        })
-    }
-
-    fn test_download_config(output_dir: std::path::PathBuf, session_id: &str) -> DownloadConfig {
-        DownloadConfig::new(
-            "https://example.com/test.flv",
-            output_dir,
-            "test-streamer-id",
-            "TestStreamer",
-            session_id,
-        )
-    }
-
-    async fn start_scripted_download(
-        manager: &DownloadManager,
-        config: DownloadConfig,
-        events: Vec<SegmentEvent>,
-    ) -> Result<String> {
-        start_scripted_download_with_engine(manager, config, ScriptedSegmentEngine::new(events))
-            .await
-    }
-
-    async fn start_scripted_download_with_engine(
-        manager: &DownloadManager,
-        config: DownloadConfig,
-        scripted: ScriptedSegmentEngine,
-    ) -> Result<String> {
-        let engine = EngineHandle {
-            engine: Arc::new(scripted),
-            engine_type: EngineType::Ffmpeg,
-            engine_key: EngineKey::global(EngineType::Ffmpeg),
-        };
-        let slot = manager
-            .acquire_slot(
-                AcquireRequest {
-                    session_id: config.session_id.clone(),
-                    streamer_id: config.streamer_id.clone(),
-                    streamer_name: config.streamer_name.clone(),
-                    engine_type: EngineType::Ffmpeg,
-                    priority: Priority::Normal,
-                },
-                CancellationToken::new(),
-            )
-            .await?;
-
-        manager.start_with_slot(slot, config, engine).await
-    }
-
-    async fn collect_segment_completed(
-        events: &mut broadcast::Receiver<DownloadManagerEvent>,
-    ) -> DownloadProgressEvent {
-        loop {
-            let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for download event")
-                .expect("download event channel closed");
-            if let DownloadManagerEvent::Progress(
-                progress @ DownloadProgressEvent::SegmentCompleted { .. },
-            ) = event
-            {
-                return progress;
-            }
-        }
-    }
-
-    async fn wait_for_download_terminal(
-        events: &mut broadcast::Receiver<DownloadManagerEvent>,
-    ) -> DownloadTerminalEvent {
-        loop {
-            let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for terminal download event")
-                .expect("download event channel closed");
-            if let DownloadManagerEvent::Terminal(terminal) = event {
-                return terminal;
-            }
-        }
-    }
-
-    async fn collect_segment_progress_events(
-        events: &mut broadcast::Receiver<DownloadManagerEvent>,
-        count: usize,
-    ) -> Vec<DownloadProgressEvent> {
-        let mut collected = Vec::new();
-        while collected.len() < count {
-            let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for download event")
-                .expect("download event channel closed");
-            if let DownloadManagerEvent::Progress(
-                progress @ (DownloadProgressEvent::SegmentStarted { .. }
-                | DownloadProgressEvent::SegmentCompleted { .. }),
-            ) = event
-            {
-                collected.push(progress);
-            }
-        }
-        collected
-    }
-
-    #[test]
-    fn test_download_manager_creation() {
-        let manager = DownloadManager::new();
-        assert_eq!(manager.active_count(), 0);
-        assert!(!manager.available_engines().is_empty());
-    }
-
-    #[tokio::test]
-    async fn download_info_views_project_consistent_live_metadata() {
-        let manager = DownloadManager::new();
-        let directory = tempfile::tempdir().unwrap();
-        let mut config = test_download_config(directory.path().to_path_buf(), "projection-session");
-        config
-            .headers
-            .push(("Authorization".to_owned(), "private-header".repeat(1024)));
-        config.engines_override =
-            Some(serde_json::json!({"nested": ["private-config".repeat(1024)]}));
-        let id = start_scripted_download_with_engine(
-            &manager,
-            config,
-            ScriptedSegmentEngine::with_shutdown_tail(vec![], vec![]),
-        )
-        .await
-        .unwrap();
-        let initial = manager.get_active_downloads().pop().unwrap();
-        assert_eq!(initial.id, id);
-        assert_eq!(initial.session_id, "projection-session");
-        let by_streamer = manager
-            .get_download_by_streamer(&initial.streamer_id)
-            .unwrap();
-        let by_status = manager
-            .get_downloads_by_status(initial.status)
-            .pop()
-            .unwrap();
-        assert_eq!(by_streamer.url, initial.url);
-        assert_eq!(by_status.url, initial.url);
-        manager
-            .active_downloads
-            .get(&id)
-            .unwrap()
-            .handle
-            .update_config(|config| {
-                config.url = "https://invalid.test/refreshed".to_owned();
-            });
-        assert_eq!(
-            manager
-                .get_download_by_streamer(&initial.streamer_id)
-                .unwrap()
-                .url,
-            "https://invalid.test/refreshed"
-        );
-        assert_ne!(
-            initial.url, "https://invalid.test/refreshed",
-            "earlier snapshots own their projected fields"
-        );
-        tokio::time::timeout(Duration::from_secs(2), manager.stop_download(&id))
-            .await
-            .unwrap()
-            .unwrap();
-        assert!(manager.get_active_downloads().is_empty());
-        assert!(
-            manager
-                .get_download_by_streamer(&initial.streamer_id)
-                .is_none()
-        );
-        assert!(manager.get_downloads_by_status(initial.status).is_empty());
-    }
-
-    #[tokio::test]
-    async fn maintenance_admission_fences_preacquired_recording_starts() {
-        let manager = DownloadManager::new();
-        let temp = tempfile::tempdir().unwrap();
-        let config = test_download_config(temp.path().to_path_buf(), "maintenance-fence");
-        let slot = manager
-            .acquire_slot(
-                AcquireRequest {
-                    session_id: config.session_id.clone(),
-                    streamer_id: config.streamer_id.clone(),
-                    streamer_name: config.streamer_name.clone(),
-                    engine_type: EngineType::Ffmpeg,
-                    priority: Priority::Normal,
-                },
-                CancellationToken::new(),
-            )
-            .await
-            .unwrap();
-        let admission = manager
-            .try_admit_maintenance(0)
-            .expect("idle maintenance admitted");
-        let engine = EngineHandle {
-            engine: Arc::new(ScriptedSegmentEngine::with_shutdown_tail(vec![], vec![])),
-            engine_type: EngineType::Ffmpeg,
-            engine_key: EngineKey::global(EngineType::Ffmpeg),
-        };
-        let mut start = Box::pin(manager.start_with_slot(slot, config, engine));
-        assert!(
-            futures::poll!(start.as_mut()).is_pending(),
-            "even a preacquired slot cannot bypass maintenance"
-        );
-        assert_eq!(manager.active_count(), 0);
-        assert!(manager.try_admit_maintenance(0).is_none());
-        drop(admission);
-        let id = tokio::time::timeout(Duration::from_secs(2), start)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(manager.active_count(), 1);
-        assert!(manager.try_admit_maintenance(0).is_none());
-        let admission = manager
-            .try_admit_maintenance(1)
-            .expect("configured active threshold is honored");
-        // Completion does not acquire the admission lock: it must remain able
-        // to settle an already running recording while maintenance is admitted.
-        tokio::time::timeout(Duration::from_secs(2), manager.stop_download(&id))
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(manager.active_count(), 0);
-        drop(admission);
-        assert!(manager.try_admit_maintenance(0).is_some());
-        let operation = manager.begin_operation().await.unwrap();
-        assert!(
-            manager.try_admit_maintenance(0).is_none(),
-            "in-flight admission work conservatively defers maintenance"
-        );
-        drop(operation);
-    }
-
-    #[tokio::test]
-    async fn maintenance_admission_respects_shutdown_and_cancelled_starts() {
-        let manager = DownloadManager::new();
-        let admission = manager.try_admit_maintenance(0).unwrap();
-        let mut operation = Box::pin(manager.begin_operation());
-        assert!(futures::poll!(operation.as_mut()).is_pending());
-        drop(operation);
-        drop(admission);
-        assert!(
-            manager.try_admit_maintenance(0).is_some(),
-            "cancelled waiter does not retain the fence"
-        );
-        let report = tokio::time::timeout(
-            Duration::from_secs(2),
-            manager.shutdown_until(tokio::time::Instant::now() + Duration::from_secs(1)),
-        )
-        .await
-        .unwrap();
-        assert!(report.failures.is_empty());
-        assert!(
-            manager.try_admit_maintenance(0).is_none(),
-            "maintenance cannot race shutdown admission"
-        );
-    }
-
-    #[tokio::test]
-    async fn cancelled_queued_start_clears_event_and_capacity_while_maintenance_is_admitted() {
-        enum CancelMode {
-            DropFuture,
-            BeforeStart,
-            DuringWait,
-        }
-        for mode in [
-            CancelMode::DropFuture,
-            CancelMode::BeforeStart,
-            CancelMode::DuringWait,
-        ] {
-            let manager = DownloadManager::with_config(DownloadManagerConfig {
-                max_concurrent_downloads: 1,
-                high_priority_extra_slots: 0,
-                ..Default::default()
-            });
-            let mut events = manager.subscribe();
-            let request = |session: &str| AcquireRequest {
-                session_id: session.to_owned(),
-                streamer_id: "streamer".to_owned(),
-                streamer_name: "Streamer".to_owned(),
-                engine_type: EngineType::Ffmpeg,
-                priority: Priority::Normal,
-            };
-            let holder = manager
-                .acquire_slot(request("holder"), CancellationToken::new())
-                .await
-                .unwrap();
-            let mut queued =
-                Box::pin(manager.acquire_slot(request("queued"), CancellationToken::new()));
-            assert!(futures::poll!(queued.as_mut()).is_pending());
-            drop(holder);
-            let slot = tokio::time::timeout(Duration::from_secs(2), queued)
-                .await
-                .unwrap()
-                .unwrap();
-            assert!(slot.queued_event_emitted());
-            let admission = manager.try_admit_maintenance(0).unwrap();
-            let config = DownloadConfig::new(
-                "https://invalid.test/live",
-                "recordings",
-                "streamer",
-                "Streamer",
-                "queued",
-            );
-            let engine = EngineHandle {
-                engine: Arc::new(ScriptedSegmentEngine::with_shutdown_tail(vec![], vec![])),
-                engine_type: EngineType::Ffmpeg,
-                engine_key: EngineKey::global(EngineType::Ffmpeg),
-            };
-            let cancel = CancellationToken::new();
-            if matches!(mode, CancelMode::BeforeStart) {
-                cancel.cancel();
-            }
-            let mut start =
-                Box::pin(manager.start_with_slot_cancellable(slot, config, engine, &cancel));
-            if !matches!(mode, CancelMode::BeforeStart) {
-                assert!(futures::poll!(start.as_mut()).is_pending());
-            }
-            match mode {
-                CancelMode::DropFuture => drop(start),
-                CancelMode::BeforeStart | CancelMode::DuringWait => {
-                    cancel.cancel();
-                    assert!(
-                        tokio::time::timeout(Duration::from_secs(2), start)
-                            .await
-                            .unwrap()
-                            .unwrap()
-                            .is_none()
-                    );
-                }
-            }
-            let mut queued_events = 0;
-            let mut dequeued_events = 0;
-            while let Ok(event) = events.try_recv() {
-                match event {
-                    DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadQueued {
-                        session_id,
-                        ..
-                    }) if session_id == "queued" => queued_events += 1,
-                    DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadDequeued {
-                        session_id,
-                        ..
-                    }) if session_id == "queued" => dequeued_events += 1,
-                    DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadStarted {
-                        ..
-                    }) => panic!("cancelled start must not launch an engine"),
-                    _ => {}
-                }
-            }
-            assert_eq!(queued_events, 1);
-            assert_eq!(dequeued_events, 1);
-            drop(admission);
-            let replacement = tokio::time::timeout(
-                Duration::from_secs(2),
-                manager.acquire_slot(request("queued"), CancellationToken::new()),
-            )
-            .await
-            .unwrap()
-            .unwrap();
-            assert!(
-                !replacement.queued_event_emitted(),
-                "cancelled start releases capacity and its session reservation"
-            );
-            assert_eq!(
-                manager.active_count(),
-                0,
-                "cancelled start stays stopped after maintenance releases admission"
-            );
-        }
-    }
-
-    #[test]
-    fn relative_segment_paths_are_resolved_without_canonicalizing() {
-        let relative = std::path::Path::new("recordings").join("segment.flv");
-        let expected = std::env::current_dir()
-            .expect("current directory")
-            .join(&relative);
-
-        assert_eq!(resolve_segment_path(&relative), expected.to_string_lossy());
-    }
-
-    /// `emit_rejected` must reach subscribers exactly like a
-    /// preflight-emitted rejection: the session lifecycle relies on the
-    /// `Rejected` terminal to close the session and the scheduler relies on
-    /// it to reschedule the streamer actor after `retry_after_secs`.
-    #[tokio::test]
-    async fn emit_rejected_broadcasts_terminal_event() {
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-
-        manager
-            .emit_rejected(
-                "streamer-1".to_string(),
-                "Streamer One".to_string(),
-                "session-1".to_string(),
-                "streamer temporarily disabled (error backoff)".to_string(),
-                Some(62),
-                DownloadRejectedKind::StreamerBackoff,
-            )
-            .await
-            .expect("rejection should publish");
-
-        let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-            .await
-            .expect("event should arrive")
-            .expect("channel open");
-
-        let DownloadManagerEvent::Terminal(DownloadTerminalEvent::Rejected {
-            streamer_id,
-            session_id,
-            retry_after_secs,
-            kind,
-            ..
-        }) = event
-        else {
-            panic!("expected Rejected terminal event, got {event:?}");
-        };
-        assert_eq!(streamer_id, "streamer-1");
-        assert_eq!(session_id, "session-1");
-        assert_eq!(retry_after_secs, Some(62));
-        assert!(matches!(kind, DownloadRejectedKind::StreamerBackoff));
-    }
-
-    #[tokio::test]
-    async fn repeated_engine_local_indices_use_session_scoped_indices() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-        let session_id = "session-reused-local-index";
-
-        let first_path = temp.path().join("segment-a.flv");
-        start_scripted_download(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), session_id),
-            vec![completed_segment(first_path, 0)],
-        )
-        .await
-        .expect("first download should start");
-        let first = collect_segment_completed(&mut events).await;
-        wait_for_download_terminal(&mut events).await;
-
-        let second_path = temp.path().join("segment-b.flv");
-        start_scripted_download(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), session_id),
-            vec![completed_segment(second_path, 0)],
-        )
-        .await
-        .expect("second download should start");
-        let second = collect_segment_completed(&mut events).await;
-        wait_for_download_terminal(&mut events).await;
-
-        let DownloadProgressEvent::SegmentCompleted {
-            segment_index: first_index,
-            ..
-        } = first
-        else {
-            panic!("expected first completed segment event");
-        };
-        let DownloadProgressEvent::SegmentCompleted {
-            segment_index: second_index,
-            ..
-        } = second
-        else {
-            panic!("expected second completed segment event");
-        };
-
-        assert_eq!(first_index, 0);
-        assert_eq!(second_index, 1);
-    }
-
-    #[tokio::test]
-    async fn segment_started_and_completed_share_session_index() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-        let segment_path = temp.path().join("segment.flv");
-
-        start_scripted_download(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), "session-start-complete"),
-            vec![
-                SegmentEvent::SegmentStarted {
-                    path: segment_path.clone(),
-                    sequence: 0,
-                    started_at: Utc::now(),
-                },
-                completed_segment(segment_path, 0),
-            ],
-        )
-        .await
-        .expect("download should start");
-
-        let segment_events = collect_segment_progress_events(&mut events, 2).await;
-        let DownloadProgressEvent::SegmentStarted {
-            segment_index: started_index,
-            ..
-        } = &segment_events[0]
-        else {
-            panic!("expected segment started event first");
-        };
-        let DownloadProgressEvent::SegmentCompleted {
-            segment_index: completed_index,
-            ..
-        } = &segment_events[1]
-        else {
-            panic!("expected segment completed event second");
-        };
-
-        assert_eq!(*started_index, 0);
-        assert_eq!(completed_index, started_index);
-        wait_for_download_terminal(&mut events).await;
-    }
-
-    #[tokio::test]
-    async fn segment_completion_preserves_engine_reported_path() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let nested = temp.path().join("nested");
-        tokio::fs::create_dir_all(&nested)
-            .await
-            .expect("create nested directory");
-        tokio::fs::write(temp.path().join("segment.flv"), b"segment")
-            .await
-            .expect("create segment");
-
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-        let segment_path = nested.join("..").join("segment.flv");
-        let expected_path = segment_path.to_string_lossy().to_string();
-
-        start_scripted_download(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), "session-path-identity"),
-            vec![
-                SegmentEvent::SegmentStarted {
-                    path: segment_path.clone(),
-                    sequence: 0,
-                    started_at: Utc::now(),
-                },
-                completed_segment(segment_path, 0),
-            ],
-        )
-        .await
-        .expect("download should start");
-
-        let segment_events = collect_segment_progress_events(&mut events, 2).await;
-        let DownloadProgressEvent::SegmentStarted {
-            segment_path: started_path,
-            ..
-        } = &segment_events[0]
-        else {
-            panic!("expected segment started event first");
-        };
-        let DownloadProgressEvent::SegmentCompleted {
-            segment_path: completed_path,
-            ..
-        } = &segment_events[1]
-        else {
-            panic!("expected segment completed event second");
-        };
-
-        assert_eq!(started_path, &expected_path);
-        assert_eq!(completed_path, &expected_path);
-
-        loop {
-            let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for terminal event")
-                .expect("download event channel closed");
-            if let DownloadManagerEvent::Terminal(DownloadTerminalEvent::Completed {
-                file_path,
-                ..
-            }) = event
-            {
-                assert_eq!(file_path.as_deref(), Some(expected_path.as_str()));
-                break;
-            }
-        }
-    }
-
-    /// Trailing `SegmentCompleted` flushed after `stop_download_with_reason`
-    /// must reuse the session_segment_index that the matching
-    /// `SegmentStarted` allocated. The mapping is the spawn loop's local
-    /// `engine_to_session` `HashMap`, which is alive as long as the loop is
-    /// draining the engine's channel and is unaffected by
-    /// `active_downloads.remove(download_id)`.
-    #[tokio::test]
-    async fn segment_completed_after_stop_keeps_session_index() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-        let segment_path = temp.path().join("segment.flv");
-        let release_tail = Arc::new(tokio::sync::Notify::new());
-
-        let scripted = ScriptedSegmentEngine::with_gated_tail(
-            vec![SegmentEvent::SegmentStarted {
-                path: segment_path.clone(),
-                sequence: 0,
-                started_at: Utc::now(),
-            }],
-            release_tail.clone(),
-            vec![SegmentEvent::SegmentCompleted(SegmentInfo {
-                path: segment_path,
-                duration_secs: 1.0,
-                size_bytes: 128,
-                index: 0,
-                started_at: None,
-                completed_at: Utc::now(),
-                split_reason_code: None,
-                split_reason_details_json: None,
-            })],
-        );
-
-        let download_id = start_scripted_download_with_engine(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), "session-trailing-flush"),
-            scripted,
-        )
-        .await
-        .expect("download should start");
-
-        // Drain the SegmentStarted so we know its session_index.
-        let started_index = loop {
-            let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for segment started")
-                .expect("download event channel closed");
-            if let DownloadManagerEvent::Progress(DownloadProgressEvent::SegmentStarted {
-                segment_index,
-                ..
-            }) = event
-            {
-                break segment_index;
-            }
-        };
-
-        // Stop must stay pending while the engine is still flushing its tail.
-        let mut stop = Box::pin(manager.stop_download(&download_id));
-        assert!(
-            tokio::time::timeout(Duration::from_millis(20), stop.as_mut())
-                .await
-                .is_err(),
-            "stop returned before the engine flushed its trailing segment"
-        );
-
-        // Release the trailing SegmentCompleted.
-        release_tail.notify_one();
-        stop.await
-            .expect("stop should finish after the tail drains");
-
-        // The trailing SegmentCompleted must report the same session_index
-        // as the SegmentStarted.
-        let completed = collect_segment_completed(&mut events).await;
-        let DownloadProgressEvent::SegmentCompleted {
-            segment_index: completed_index,
-            ..
-        } = completed
-        else {
-            panic!("expected segment completed event");
-        };
-        assert_eq!(completed_index, started_index);
-
-        let terminal = wait_for_download_terminal(&mut events).await;
-        assert!(matches!(
-            terminal,
-            DownloadTerminalEvent::Completed {
-                stop_cause: Some(DownloadStopCause::User),
-                ..
-            }
-        ));
-    }
-
-    #[tokio::test]
-    async fn shutdown_waits_for_trailing_segment_and_emits_one_terminal_outcome() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let (coordination_tx, mut coordination_rx) = download_coordination_channel();
-        let manager = DownloadManager::new().with_coordination_sender(coordination_tx);
-        let mut events = manager.subscribe();
-        let segment_path = temp.path().join("shutdown-tail.flv");
-        let completion_received = Arc::new(tokio::sync::Notify::new());
-        let release_completion = Arc::new(tokio::sync::Notify::new());
-        let coordination_completion_received = completion_received.clone();
-        let coordination_release_completion = release_completion.clone();
-        let coordination_task = tokio::spawn(async move {
-            while let Some(delivery) = coordination_rx
-                .recv()
-                .await
-                .expect("coordination channel should close with a marker")
-            {
-                let (event, acknowledgement) = delivery.into_parts();
-                if matches!(
-                    event,
-                    DownloadManagerEvent::Progress(DownloadProgressEvent::SegmentCompleted { .. })
-                ) {
-                    coordination_completion_received.notify_one();
-                    coordination_release_completion.notified().await;
-                }
-                acknowledgement
-                    .send(Ok(()))
-                    .expect("attempt should await its acknowledgement");
-            }
-        });
-
-        let scripted = ScriptedSegmentEngine::with_shutdown_tail(
-            vec![SegmentEvent::SegmentStarted {
-                path: segment_path.clone(),
-                sequence: 0,
-                started_at: Utc::now(),
-            }],
-            vec![completed_segment(segment_path, 0)],
-        );
-
-        start_scripted_download_with_engine(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), "session-shutdown-tail"),
-            scripted,
-        )
-        .await
-        .expect("download should start");
-
-        loop {
-            let event = tokio::time::timeout(Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for segment start")
-                .expect("download event channel closed");
-            if matches!(
-                event,
-                DownloadManagerEvent::Progress(DownloadProgressEvent::SegmentStarted { .. })
-            ) {
-                break;
-            }
-        }
-
-        let shutdown = manager.shutdown_until(tokio::time::Instant::now() + Duration::from_secs(1));
-        tokio::pin!(shutdown);
-        tokio::select! {
-            () = completion_received.notified() => {}
-            result = &mut shutdown => panic!(
-                "shutdown returned before SegmentCompleted acknowledgement: {result:?}"
-            ),
-        }
-        assert!(
-            tokio::time::timeout(Duration::from_millis(20), shutdown.as_mut())
-                .await
-                .is_err(),
-            "shutdown must remain pending until SegmentCompleted is durably applied"
-        );
-        release_completion.notify_one();
-        shutdown.await;
-        coordination_task
-            .await
-            .expect("coordination task should finish after the drain marker");
-
-        let mut saw_segment_completed = false;
-        let terminal = loop {
-            let event = tokio::time::timeout(Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for shutdown events")
-                .expect("download event channel closed");
-            match event {
-                DownloadManagerEvent::Progress(DownloadProgressEvent::SegmentCompleted {
-                    ..
-                }) => saw_segment_completed = true,
-                DownloadManagerEvent::Terminal(terminal) => break terminal,
-                _ => {}
-            }
-        };
-
-        assert!(
-            saw_segment_completed,
-            "the final segment must be published before the terminal outcome"
-        );
-        assert!(matches!(
-            terminal,
-            DownloadTerminalEvent::Cancelled {
-                cause: DownloadStopCause::Shutdown,
-                ..
-            }
-        ));
-        let mut additional_terminals = 0;
-        while let Ok(Ok(event)) =
-            tokio::time::timeout(Duration::from_millis(20), events.recv()).await
-        {
-            if matches!(event, DownloadManagerEvent::Terminal(_)) {
-                additional_terminals += 1;
-            }
-        }
-        assert_eq!(additional_terminals, 0, "terminal outcome must be unique");
-        assert_eq!(manager.active_count(), 0);
-    }
-
-    #[tokio::test]
-    async fn shutdown_rejects_queued_start_before_active_slot_is_released() {
-        let temp = tempfile::tempdir().expect("temp directory should be created");
-        let manager = Arc::new(DownloadManager::with_config(DownloadManagerConfig {
-            max_concurrent_downloads: 1,
-            high_priority_extra_slots: 0,
-            ..Default::default()
-        }));
-        manager.register_engine(Arc::new(ScriptedSegmentEngine::with_shutdown_tail(
-            Vec::new(),
-            Vec::new(),
-        )));
-        let mut events = manager.subscribe();
-
-        manager
-            .start_download(
-                DownloadConfig::new(
-                    "https://example.com/active.flv",
-                    temp.path().to_path_buf(),
-                    "active-streamer",
-                    "Active Streamer",
-                    "active-session",
-                ),
-                Some("ffmpeg".to_string()),
-                false,
-            )
-            .await
-            .expect("first download should occupy the only slot");
-
-        let queued_manager = manager.clone();
-        let queued_output = temp.path().to_path_buf();
-        let queued = tokio::spawn(async move {
-            queued_manager
-                .start_download(
-                    DownloadConfig::new(
-                        "https://example.com/queued.flv",
-                        queued_output,
-                        "queued-streamer",
-                        "Queued Streamer",
-                        "queued-session",
-                    ),
-                    Some("ffmpeg".to_string()),
-                    false,
-                )
-                .await
-        });
-
-        tokio::time::timeout(Duration::from_secs(1), async {
-            while manager.queue.pending_count() != 1 {
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("second download should park in the queue");
-
-        let report = manager
-            .shutdown_until(tokio::time::Instant::now() + Duration::from_secs(1))
-            .await;
-        let queued_error = queued
-            .await
-            .expect("queued start task should not panic")
-            .expect_err("queued start must be rejected during shutdown");
-
-        assert!(queued_error.to_string().contains("shutting down"));
-        assert_eq!(report.stopped_download_ids.len(), 1);
-        assert!(report.deadline_exceeded_download_ids.is_empty());
-        assert_eq!(manager.active_count(), 0);
-        assert_eq!(manager.queue.pending_count(), 0);
-
-        let started_count = std::iter::from_fn(|| events.try_recv().ok())
-            .filter(|event| {
-                matches!(
-                    event,
-                    DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadStarted { .. })
-                )
-            })
-            .count();
-        assert_eq!(started_count, 1, "the queued engine must never start");
-    }
-
-    /// `clear_session_segment_index` can fire mid-drain (the dedicated
-    /// `SessionTransition::Ended` subscriber in `services/container.rs` runs
-    /// as soon as `SessionLifecycle` writes `end_time`, independent of the
-    /// download's spawn loop). The trailing `SegmentCompleted` that follows
-    /// must still report the same session_index as its matching
-    /// `SegmentStarted`. The spawn loop's local `engine_to_session` map is
-    /// what makes this hold — it survives the session-counter eviction.
-    ///
-    /// The prelude includes a complete first segment (Start+Complete for
-    /// engine seq=0) so the session counter advances past 0 before the
-    /// second `SegmentStarted` (seq=1) lands. A naive design that wipes the
-    /// mapping along with the counter would re-allocate the trailing
-    /// `SegmentCompleted(1)` as session_index=0 from the recreated counter,
-    /// not 1 — that's what this test rules out.
-    #[tokio::test]
-    async fn segment_completed_after_session_cleanup_keeps_index() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-        let first_path = temp.path().join("segment-0.flv");
-        let second_path = temp.path().join("segment-1.flv");
-        let release_tail = Arc::new(tokio::sync::Notify::new());
-        let session_id = "session-cleanup-mid-drain";
-
-        let scripted = ScriptedSegmentEngine::with_gated_tail(
-            vec![
-                SegmentEvent::SegmentStarted {
-                    path: first_path.clone(),
-                    sequence: 0,
-                    started_at: Utc::now(),
-                },
-                completed_segment(first_path, 0),
-                SegmentEvent::SegmentStarted {
-                    path: second_path.clone(),
-                    sequence: 1,
-                    started_at: Utc::now(),
-                },
-            ],
-            release_tail.clone(),
-            vec![completed_segment(second_path.clone(), 1)],
-        );
-
-        start_scripted_download_with_engine(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), session_id),
-            scripted,
-        )
-        .await
-        .expect("download should start");
-
-        // Drain events until we see the SegmentStarted for engine seq=1 —
-        // that's the one whose matching SegmentCompleted is gated.
-        let started_index = loop {
-            let event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for segment started")
-                .expect("download event channel closed");
-            if let DownloadManagerEvent::Progress(DownloadProgressEvent::SegmentStarted {
-                segment_index,
-                segment_path,
-                ..
-            }) = event
-                && segment_path == second_path.to_string_lossy()
-            {
-                break segment_index;
-            }
-        };
-        // With the counter advanced past 0, the second SegmentStarted's
-        // session_index must be > 0; otherwise the trailing assertion below
-        // can't distinguish fixed from broken.
-        assert!(
-            started_index > 0,
-            "test setup expects session counter to advance past 0 before the gated tail; got {started_index}"
-        );
-
-        // Simulate the `SessionTransition::Ended` subscriber firing while the
-        // engine is still parked.
-        manager.clear_session_segment_index(session_id);
-
-        // Release the trailing SegmentCompleted.
-        release_tail.notify_one();
-
-        let completed = collect_segment_completed(&mut events).await;
-        let DownloadProgressEvent::SegmentCompleted {
-            segment_index: completed_index,
-            ..
-        } = completed
-        else {
-            panic!("expected segment completed event");
-        };
-        assert_eq!(completed_index, started_index);
-    }
-
-    /// `shutdown_until` contains attempts past its deadline, so an engine that
-    /// ignores `DownloadHandle::cancel` keeps it pending forever. A caller that
-    /// stops awaiting it must still be able to reclaim those attempts:
-    /// `abort_attempts` drops each attempt future, which is what kills the
-    /// engine child owned through `kill_on_drop`.
-    #[tokio::test]
-    async fn abort_attempts_reclaims_attempts_left_by_a_dropped_shutdown() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let manager = DownloadManager::new();
-        let mut events = manager.subscribe();
-        let segment_path = temp.path().join("segment-0.flv");
-        // Never notified, so the engine parks past the shutdown request.
-        let wedged = Arc::new(tokio::sync::Notify::new());
-
-        let scripted = ScriptedSegmentEngine::with_gated_tail(
-            vec![SegmentEvent::SegmentStarted {
-                path: segment_path.clone(),
-                sequence: 0,
-                started_at: Utc::now(),
-            }],
-            wedged,
-            Vec::new(),
-        );
-        let download_id = start_scripted_download_with_engine(
-            &manager,
-            test_download_config(temp.path().to_path_buf(), "session-abort-attempts"),
-            scripted,
-        )
-        .await
-        .expect("download should start");
-
-        loop {
-            let event = tokio::time::timeout(Duration::from_secs(1), events.recv())
-                .await
-                .expect("timed out waiting for segment started")
-                .expect("download event channel closed");
-            if matches!(
-                event,
-                DownloadManagerEvent::Progress(DownloadProgressEvent::SegmentStarted { .. })
-            ) {
-                break;
-            }
-        }
-
-        assert!(
-            tokio::time::timeout(
-                Duration::from_millis(50),
-                manager.shutdown_until(tokio::time::Instant::now()),
-            )
-            .await
-            .is_err(),
-            "shutdown_until must keep containing an attempt whose engine ignores cancellation"
-        );
-
-        let aborted = manager
-            .abort_attempts(tokio::time::Instant::now() + Duration::from_secs(5))
-            .await;
-        assert_eq!(aborted, vec![download_id]);
-        assert_eq!(manager.active_count(), 0);
-    }
-
-    #[tokio::test]
-    async fn test_runtime_reconfigure_max_concurrent_downloads() {
-        // Validates the public contract: increasing capacity is
-        // immediately observable, decreasing capacity is observable
-        // even though existing in-flight downloads aren't preempted.
-        let config = DownloadManagerConfig {
-            max_concurrent_downloads: 2,
-            high_priority_extra_slots: 0,
-            ..Default::default()
-        };
-        let manager = DownloadManager::with_config(config);
-
-        // Increase beyond initial capacity.
-        assert_eq!(manager.set_max_concurrent_downloads(4), 4);
-        assert_eq!(manager.max_concurrent_downloads(), 4);
-
-        // Decrease — getter reflects the new value immediately.
-        assert_eq!(manager.set_max_concurrent_downloads(1), 1);
-        assert_eq!(manager.max_concurrent_downloads(), 1);
-
-        // After saturating, a third acquire queues. We verify by
-        // calling acquire on the queue directly — this mirrors the
-        // semaphore-probe assertions the previous version made, but
-        // through the public abstraction.
-        let q = manager.queue.clone();
-        let req = AcquireRequest {
-            session_id: "s1".to_string(),
-            streamer_id: "x".to_string(),
-            streamer_name: "x".to_string(),
-            engine_type: EngineType::Ffmpeg,
-            priority: Priority::Normal,
-        };
-        let _slot1 = q
-            .acquire(req, CancellationToken::new(), |_| {})
-            .await
-            .unwrap();
-        assert_eq!(q.in_flight(), 1);
-
-        // Second acquire must queue (capacity = 1).
-        let q2 = q.clone();
-        let h = tokio::spawn(async move {
-            q2.acquire(
-                AcquireRequest {
-                    session_id: "s2".to_string(),
-                    streamer_id: "x".to_string(),
-                    streamer_name: "x".to_string(),
-                    engine_type: EngineType::Ffmpeg,
-                    priority: Priority::Normal,
-                },
-                CancellationToken::new(),
-                |_| {},
-            )
-            .await
-        });
-        // Wait for it to register as pending.
-        for _ in 0..50 {
-            if q.pending_count() == 1 {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(5)).await;
-        }
-        assert_eq!(q.pending_count(), 1);
-
-        // Bump the limit; the waiter should fire.
-        manager.set_max_concurrent_downloads(2);
-        let _slot2 = h.await.unwrap().unwrap();
-        assert_eq!(q.in_flight(), 2);
-    }
-
-    #[tokio::test]
-    async fn queued_slot_abandoned_after_acquire_emits_dequeued() {
-        let config = DownloadManagerConfig {
-            max_concurrent_downloads: 1,
-            high_priority_extra_slots: 0,
-            ..Default::default()
-        };
-        let manager = Arc::new(DownloadManager::with_config(config));
-        let mut events = manager.subscribe();
-
-        let first = manager
-            .acquire_slot(
-                AcquireRequest {
-                    session_id: "active-session".to_string(),
-                    streamer_id: "streamer-active".to_string(),
-                    streamer_name: "Active".to_string(),
-                    engine_type: EngineType::Ffmpeg,
-                    priority: Priority::Normal,
-                },
-                CancellationToken::new(),
-            )
-            .await
-            .unwrap();
-
-        let waiter_manager = manager.clone();
-        let waiter = tokio::spawn(async move {
-            waiter_manager
-                .acquire_slot(
-                    AcquireRequest {
-                        session_id: "queued-session".to_string(),
-                        streamer_id: "streamer-queued".to_string(),
-                        streamer_name: "Queued".to_string(),
-                        engine_type: EngineType::Ffmpeg,
-                        priority: Priority::Normal,
-                    },
-                    CancellationToken::new(),
-                )
-                .await
-                .unwrap()
-        });
-
-        let queued_event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-            .await
-            .unwrap()
-            .unwrap();
-        assert!(matches!(
-            queued_event,
-            DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadQueued {
-                ref session_id,
-                ..
-            }) if session_id == "queued-session"
-        ));
-
-        drop(first);
-        let slot = waiter.await.unwrap();
-        assert!(slot.queued_event_emitted());
-
-        manager.emit_dequeued_for_slot(&slot, "streamer-queued", "Queued");
-
-        let dequeued_event = tokio::time::timeout(std::time::Duration::from_secs(1), events.recv())
-            .await
-            .unwrap()
-            .unwrap();
-        assert!(matches!(
-            dequeued_event,
-            DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadDequeued {
-                ref session_id,
-                ..
-            }) if session_id == "queued-session"
-        ));
-    }
-
-    #[test]
-    fn test_engine_registration() {
-        let manager = DownloadManager::new();
-
-        // FFmpeg should be registered by default
-        assert!(manager.get_engine(EngineType::Ffmpeg).is_some());
-        assert!(manager.get_engine(EngineType::Streamlink).is_some());
-        assert!(manager.get_engine(EngineType::Mesio).is_some());
-    }
-
-    #[test]
-    fn test_determine_config_update_type_cookies_only() {
-        let update = PendingConfigUpdate::new(Some("session=abc123".to_string()), None, None);
-        assert_eq!(
-            DownloadManager::determine_config_update_type(&update),
-            ConfigUpdateType::Cookies
-        );
-    }
-
-    #[test]
-    fn test_determine_config_update_type_headers_only() {
-        let update = PendingConfigUpdate::new(
-            None,
-            Some(vec![(
-                "Authorization".to_string(),
-                "Bearer token".to_string(),
-            )]),
-            None,
-        );
-        assert_eq!(
-            DownloadManager::determine_config_update_type(&update),
-            ConfigUpdateType::Headers
-        );
-    }
-
-    #[test]
-    fn test_determine_config_update_type_retry_only() {
-        let update = PendingConfigUpdate::new(None, None, Some(RetryConfig::default()));
-        assert_eq!(
-            DownloadManager::determine_config_update_type(&update),
-            ConfigUpdateType::RetryConfig
-        );
-    }
-
-    #[test]
-    fn test_determine_config_update_type_multiple() {
-        let update = PendingConfigUpdate::new(
-            Some("session=abc123".to_string()),
-            Some(vec![(
-                "Authorization".to_string(),
-                "Bearer token".to_string(),
-            )]),
-            None,
-        );
-        assert_eq!(
-            DownloadManager::determine_config_update_type(&update),
-            ConfigUpdateType::Multiple
-        );
-    }
-
-    #[test]
-    fn test_determine_config_update_type_all_three() {
-        let update = PendingConfigUpdate::new(
-            Some("session=abc123".to_string()),
-            Some(vec![(
-                "Authorization".to_string(),
-                "Bearer token".to_string(),
-            )]),
-            Some(RetryConfig::default()),
-        );
-        assert_eq!(
-            DownloadManager::determine_config_update_type(&update),
-            ConfigUpdateType::Multiple
-        );
-    }
-
-    #[test]
-    fn test_emit_config_updated_with_subscriber() {
-        let manager = DownloadManager::new();
-        let mut receiver = manager.subscribe();
-
-        let update = PendingConfigUpdate::new(Some("session=abc123".to_string()), None, None);
-
-        let result = manager.emit_config_updated("download-123", "streamer-456", &update);
-        assert!(result);
-
-        // Verify the event was received
-        let event = receiver.try_recv().unwrap();
-        match event {
-            DownloadManagerEvent::Progress(DownloadProgressEvent::ConfigUpdated {
-                download_id,
-                streamer_id,
-                update_type,
-                ..
-            }) => {
-                assert_eq!(download_id, "download-123");
-                assert_eq!(streamer_id, "streamer-456");
-                assert_eq!(update_type, ConfigUpdateType::Cookies);
-            }
-            _ => panic!("Expected ConfigUpdated event"),
-        }
-    }
-
-    #[test]
-    fn test_emit_config_updated_no_updates() {
-        let manager = DownloadManager::new();
-        let _receiver = manager.subscribe();
-
-        let update = PendingConfigUpdate::default();
-        assert!(!update.has_updates());
-
-        let result = manager.emit_config_updated("download-123", "streamer-456", &update);
-        assert!(!result);
-    }
-
-    #[test]
-    fn test_emit_config_update_failed_with_subscriber() {
-        let manager = DownloadManager::new();
-        let mut receiver = manager.subscribe();
-
-        let result =
-            manager.emit_config_update_failed("download-123", "streamer-456", "Connection timeout");
-        assert!(result);
-
-        // Verify the event was received
-        let event = receiver.try_recv().unwrap();
-        match event {
-            DownloadManagerEvent::Progress(DownloadProgressEvent::ConfigUpdateFailed {
-                download_id,
-                streamer_id,
-                error,
-                ..
-            }) => {
-                assert_eq!(download_id, "download-123");
-                assert_eq!(streamer_id, "streamer-456");
-                assert_eq!(error, "Connection timeout");
-            }
-            _ => panic!("Expected ConfigUpdateFailed event"),
-        }
-    }
-
-    // ========== Output-root write gate integration ==========
-
-    /// Build a `DownloadConfig` pointed at `output_dir`, with the other
-    /// fields set to minimal plausible values. The URL/streamer fields are
-    /// only used for logging — we never actually spawn an engine in these
-    /// tests.
-    fn test_config_with_output_dir(output_dir: std::path::PathBuf) -> DownloadConfig {
-        DownloadConfig::new(
-            "https://example.com/test.flv",
-            output_dir,
-            "test-streamer-id",
-            "TestStreamer",
-            "test-session-id",
-        )
-    }
-
-    /// Wrap a `DownloadManager` with a freshly constructed gate and return
-    /// both the manager and a counter the recovery hook bumps each time it
-    /// fires. Used by the three `prepare_output_dir_*` tests below.
-    fn manager_with_gate() -> (
-        DownloadManager,
-        Arc<std::sync::atomic::AtomicUsize>,
-        Arc<super::super::output_root_gate::OutputRootGate>,
-    ) {
-        use super::super::output_root_gate::{OutputRootGate, RecoveryHook};
-        use std::sync::Weak;
-        use std::sync::atomic::AtomicUsize;
-
-        let counter = Arc::new(AtomicUsize::new(0));
-        let c2 = counter.clone();
-        let hook: RecoveryHook = Arc::new(move |_root: &std::path::Path| {
-            c2.fetch_add(1, Ordering::SeqCst);
-        });
-        let gate = OutputRootGate::new(
-            Weak::new(),
-            hook,
-            vec![],
-            Duration::from_secs(1), // short cooldown for the test runner
-        );
-        let manager = DownloadManager::new();
-        manager.set_output_root_gate(gate.clone());
-        (manager, counter, gate)
-    }
-
-    #[tokio::test]
-    async fn prepare_output_dir_happy_path_returns_ok() {
-        // Baseline: a real, writable temp dir. The gate starts Healthy and
-        // stays Healthy; `ensure_output_dir` creates the nested subdir that
-        // doesn't yet exist inside the temp root.
-        let temp = tempfile::tempdir().expect("tempdir");
-        let nested = temp.path().join("huya").join("X").join("20260415");
-        let (manager, counter, gate) = manager_with_gate();
-
-        let config = test_config_with_output_dir(nested.clone());
-        let result = manager.prepare_output_dir(&config).await;
-
-        assert!(result.is_ok(), "happy path should succeed: {:?}", result);
-        assert!(
-            nested.is_dir(),
-            "nested output dir should have been created"
-        );
-        // Recovery hook only fires on a Degraded → Healthy transition.
-        // A first-ever success against an untracked root is a no-op for the
-        // gate, so the counter stays at zero.
-        assert_eq!(counter.load(Ordering::SeqCst), 0);
-        // And the gate snapshot should be empty (no roots ever tracked).
-        assert!(gate.snapshot().is_empty());
-    }
-
-    #[tokio::test]
-    async fn mid_stream_output_io_degrades_the_gate_before_the_terminal() {
-        for io_kind in [
-            IoErrorKindSer::StorageFull,
-            IoErrorKindSer::PermissionDenied,
-            IoErrorKindSer::ReadOnlyFilesystem,
-            IoErrorKindSer::NotFound,
-            IoErrorKindSer::TimedOut,
-        ] {
-            let temp = tempfile::tempdir().expect("tempdir");
-            let output_dir = temp.path().join("recording");
-            let (manager, _, gate) = manager_with_gate();
-            let mut events = manager.subscribe();
-            start_scripted_download(
-                &manager,
-                test_download_config(output_dir.clone(), "output-io-session"),
-                vec![
-                    SegmentEvent::OutputIoError {
-                        output_dir: output_dir.clone(),
-                        io_kind,
-                        detail: "output fault".to_owned(),
-                    },
-                    SegmentEvent::DownloadFailed {
-                        kind: DownloadFailureKind::OutputRootUnavailable { io_kind },
-                        message: "writer failed".to_owned(),
-                    },
-                ],
-            )
-            .await
-            .unwrap();
-            let terminal = wait_for_download_terminal(&mut events).await;
-            assert!(
-                matches!(terminal, DownloadTerminalEvent::Failed { kind: DownloadFailureKind::OutputRootUnavailable { io_kind: actual }, .. } if actual == io_kind)
-            );
-            let blocked = gate
-                .check(&output_dir)
-                .expect_err("gate must reject before the terminal is observed");
-            assert_eq!(blocked.kind, io_kind);
-            assert_eq!(blocked.message, "output fault");
-        }
-    }
-
-    #[tokio::test]
-    async fn output_failures_without_a_gate_record_keep_the_circuit_breaker() {
-        for attach_gate in [false, true] {
-            let temp = tempfile::tempdir().unwrap();
-            let output_dir = temp.path().join("recording");
-            let (manager, gate) = if attach_gate {
-                let (manager, _, gate) = manager_with_gate();
-                (manager, Some(gate))
-            } else {
-                (DownloadManager::new(), None)
-            };
-            let mut events = manager.subscribe();
-            let output_error = SegmentEvent::OutputIoError {
-                output_dir: output_dir.clone(),
-                io_kind: IoErrorKindSer::PermissionDenied,
-                detail: "denied".to_owned(),
-            };
-            let terminal = SegmentEvent::DownloadFailed {
-                kind: DownloadFailureKind::OutputRootUnavailable {
-                    io_kind: IoErrorKindSer::PermissionDenied,
-                },
-                message: "writer failed".to_owned(),
-            };
-            // Either the event is correctly ordered but there is no gate, or
-            // the gate is attached but the engine sent its terminal too early.
-            let scripted = if attach_gate {
-                vec![terminal, output_error]
-            } else {
-                vec![output_error, terminal]
-            };
-            start_scripted_download(
-                &manager,
-                test_download_config(output_dir.clone(), "unpaired-output-failure"),
-                scripted,
-            )
-            .await
-            .unwrap();
-            let terminal = wait_for_download_terminal(&mut events).await;
-            assert!(matches!(
-                terminal,
-                DownloadTerminalEvent::Failed {
-                    kind: DownloadFailureKind::Io,
-                    ..
-                }
-            ));
-            if let Some(gate) = gate {
-                assert!(gate.check(&output_dir).is_ok());
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn early_cancellation_does_not_charge_breaker_but_racing_failure_does() {
-        for kind in [DownloadFailureKind::Cancelled, DownloadFailureKind::Other] {
-            let manager = DownloadManager::new();
-            let temp = tempfile::tempdir().unwrap();
-            let mut events = manager.subscribe();
-            let key = EngineKey::global(EngineType::Ffmpeg);
-            for attempt in 0..5 {
-                let scripted = ScriptedSegmentEngine::with_shutdown_tail(
-                    vec![],
-                    vec![SegmentEvent::DownloadFailed {
-                        kind,
-                        message: "early engine outcome".to_owned(),
-                    }],
-                );
-                let id = start_scripted_download_with_engine(
-                    &manager,
-                    test_download_config(
-                        temp.path().to_path_buf(),
-                        &format!("early-stop-{attempt}"),
-                    ),
-                    scripted,
-                )
-                .await
-                .unwrap();
-                manager
-                    .request_stop(&id, DownloadStopCause::Shutdown)
-                    .unwrap();
-                assert!(matches!(
-                    wait_for_download_terminal(&mut events).await,
-                    DownloadTerminalEvent::Cancelled { .. }
-                ));
-            }
-            assert_eq!(
-                manager.circuit_breakers.is_allowed(&key),
-                kind == DownloadFailureKind::Cancelled
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn prepare_output_dir_on_unwritable_parent_trips_gate() {
-        // Force create_dir_all to fail portably. GHA's Windows runner
-        // runs as admin, so `C:\nonexistent\...` is creatable there;
-        // Linux CI sandboxes often mount `/` read-only → EROFS. Instead,
-        // put a regular file in a tempdir and point the output path at
-        // a child of it — `create_dir_all` then fails with NotADirectory
-        // on Unix and ERROR_DIRECTORY on Windows, both of which the gate
-        // classifies into DownloadFailureKind::OutputRootUnavailable.
-        // We don't pin the exact io_kind because the classification
-        // bucket (never `Other`) is what's actually under test.
-        let temp = tempfile::tempdir().expect("tempdir");
-        let blocker = temp.path().join("blocker");
-        std::fs::write(&blocker, b"i am a file, not a dir").unwrap();
-        let bad_output = blocker.join("huya").join("X").join("20260415");
-        let (manager, counter, gate) = manager_with_gate();
-
-        let config = test_config_with_output_dir(bad_output.clone());
-        let result = manager.prepare_output_dir(&config).await;
-
-        // Must fail with OutputRootUnavailable{...}, NOT the generic Io
-        // kind — this proves `EngineStartError::from` correctly walked
-        // the error chain and the manager's `prepare_output_dir` routed
-        // the io::Error into the gate before returning.
-        let err = result.expect_err("should fail on unwritable parent");
-        let io_kind = match err.kind {
-            DownloadFailureKind::OutputRootUnavailable { io_kind } => io_kind,
-            other => panic!("expected OutputRootUnavailable{{_}}, got {:?}", other),
-        };
-        assert!(
-            !matches!(io_kind, IoErrorKindSer::Other),
-            "expected a classified io_kind (NotFound / PermissionDenied / \
-             ReadOnlyFilesystem / StorageFull / TimedOut), got Other"
-        );
-
-        // The gate must now be tracking a Degraded root with a matching
-        // cached error kind.
-        let snapshot = gate.snapshot();
-        assert_eq!(snapshot.len(), 1);
-        assert_eq!(
-            snapshot[0].state,
-            crate::downloader::RootHealthState::Degraded
-        );
-        let last_error = snapshot[0]
-            .last_error
-            .as_ref()
-            .expect("degraded root must have cached error");
-        assert_eq!(last_error.0, io_kind);
-
-        // A second `prepare_output_dir` call inside the cooldown window
-        // must fast-reject via `gate.check()`, not re-try `create_dir_all`.
-        // This is the key property that stops the 508 cascade.
-        let result2 = manager.prepare_output_dir(&config).await;
-        let err2 = result2.expect_err("second call should fast-reject");
-        assert!(matches!(
-            err2.kind,
-            DownloadFailureKind::OutputRootUnavailable { .. }
-        ));
-        // Recovery hook must NOT have fired — we're still Degraded.
-        assert_eq!(counter.load(Ordering::SeqCst), 0);
-    }
-
-    #[tokio::test]
-    async fn prepare_output_dir_recovers_after_path_becomes_valid() {
-        // Trip the gate by pointing at a child of a regular file (fails
-        // portably on Unix ENOTDIR and Windows ERROR_DIRECTORY — tokio's
-        // create_dir_all will not recreate a file-as-directory ancestor).
-        // Then fix the filesystem, wait past the cooldown, retry. The
-        // winning CAS caller should see ensure_output_dir succeed, flip
-        // the gate to Healthy, and fire the recovery hook exactly once.
-        let temp = tempfile::tempdir().expect("tempdir");
-
-        use super::super::output_root_gate::{OutputRootGate, RecoveryHook};
-        use std::sync::Weak;
-        use std::sync::atomic::AtomicUsize;
-
-        let counter = Arc::new(AtomicUsize::new(0));
-        let c2 = counter.clone();
-        let hook: RecoveryHook = Arc::new(move |_root: &std::path::Path| {
-            c2.fetch_add(1, Ordering::SeqCst);
-        });
-        let configured_root = temp.path().to_path_buf();
-        let gate = OutputRootGate::new(
-            Weak::new(),
-            hook,
-            vec![configured_root.clone()],
-            Duration::from_secs(1),
-        );
-        let manager = DownloadManager::new();
-        manager.set_output_root_gate(gate.clone());
-
-        // `doomed` is a regular file; any path under it will fail
-        // create_dir_all because an ancestor component is not a directory.
-        let doomed = temp.path().join("doomed");
-        std::fs::write(&doomed, b"i am a file, not a dir").unwrap();
-        let under_doomed = doomed.join("will-fail");
-
-        // Now `create_dir_all(under_doomed)` will fail with NotADirectory
-        // or similar because one of the ancestor components is a regular
-        // file. On Linux this surfaces as ErrorKind::NotFound or
-        // ErrorKind::NotADirectory depending on kernel version; either
-        // way the gate records a failure.
-        let bad_config = test_config_with_output_dir(under_doomed.clone());
-        let first = manager.prepare_output_dir(&bad_config).await;
-        assert!(first.is_err(), "should fail when ancestor is a file");
-        let snap = gate.snapshot();
-        assert_eq!(snap.len(), 1, "gate should be tracking the configured root");
-        assert_eq!(snap[0].state, crate::downloader::RootHealthState::Degraded);
-        assert_eq!(counter.load(Ordering::SeqCst), 0);
-
-        // Fix the filesystem: remove the blocking file and recreate the
-        // directory structure. Wait past the 1s cooldown before retrying.
-        std::fs::remove_file(&doomed).unwrap();
-        std::fs::create_dir_all(&under_doomed).unwrap();
-        tokio::time::sleep(Duration::from_millis(1200)).await;
-
-        // Retry. The winning CAS caller's ensure_output_dir succeeds, the
-        // gate flips to Healthy, the recovery hook fires.
-        let second = manager.prepare_output_dir(&bad_config).await;
-        assert!(
-            second.is_ok(),
-            "retry should succeed after cooldown and filesystem fix: {:?}",
-            second
-        );
-
-        // Recovery hook is spawned on a tokio task; yield so it runs.
-        tokio::task::yield_now().await;
-        tokio::time::sleep(Duration::from_millis(50)).await;
-        assert_eq!(
-            counter.load(Ordering::SeqCst),
-            1,
-            "recovery hook should fire exactly once per Degraded→Healthy transition"
-        );
-
-        // Gate snapshot should now show Healthy.
-        let snap = gate.snapshot();
-        assert_eq!(snap.len(), 1);
-        assert_eq!(snap[0].state, crate::downloader::RootHealthState::Healthy);
-    }
-
-    #[tokio::test]
-    async fn prepare_output_dir_without_gate_is_transparent() {
-        // Safety guarantee: with no gate installed, prepare_output_dir is a
-        // plain create-dir call — success creates the dir, failure returns
-        // a classified EngineStartError, no panics, no hidden state.
-        let temp = tempfile::tempdir().expect("tempdir");
-        let nested = temp.path().join("a").join("b").join("c");
-        let manager = DownloadManager::new();
-        // Deliberately NO set_output_root_gate.
-
-        let config = test_config_with_output_dir(nested.clone());
-        assert!(manager.prepare_output_dir(&config).await.is_ok());
-        assert!(nested.is_dir());
-    }
-
-    #[tokio::test]
-    async fn required_terminal_event_survives_lagged_observer() {
-        let (coordination_tx, mut coordination_rx) = download_coordination_channel();
-        let manager = DownloadManager::new().with_coordination_sender(coordination_tx);
-        let mut observer = manager.subscribe();
-
-        for index in 0..300 {
-            manager.events.publish(DownloadManagerEvent::Progress(
-                DownloadProgressEvent::DownloadDequeued {
-                    streamer_id: "streamer".to_string(),
-                    streamer_name: "Streamer".to_string(),
-                    session_id: format!("session-{index}"),
-                },
-            ));
-        }
-
-        let publish = manager.emit_rejected(
-            "streamer".to_string(),
-            "Streamer".to_string(),
-            "required-session".to_string(),
-            "test rejection".to_string(),
-            Some(5),
-            DownloadRejectedKind::CircuitBreaker,
-        );
-        tokio::pin!(publish);
-
-        let delivery = tokio::select! {
-            delivery = coordination_rx.recv() => delivery
-                .expect("required terminal channel stays open")
-                .expect("required terminal event should arrive"),
-            result = &mut publish => panic!("publish returned before acknowledgement: {result:?}"),
-        };
-        let (event, acknowledgement) = delivery.into_parts();
-        let DownloadManagerEvent::Terminal(terminal) = event else {
-            panic!("expected terminal coordination event");
-        };
-        let _ = acknowledgement.send(Ok(()));
-        publish
-            .await
-            .expect("rejection should complete after acknowledgement");
-        assert_eq!(terminal.session_id(), "required-session");
-        assert!(matches!(
-            observer.recv().await,
-            Err(broadcast::error::RecvError::Lagged(_))
-        ));
-    }
-
-    #[tokio::test]
-    async fn coordination_shutdown_rejects_events_behind_marker() {
-        let (coordination_tx, mut coordination_rx) = download_coordination_channel();
-        let late_sender = coordination_tx.clone();
-        let shutdown = tokio::spawn(async move { coordination_tx.shutdown().await });
-
-        assert!(matches!(coordination_rx.recv().await, Ok(None)));
-        shutdown
-            .await
-            .expect("shutdown task should not panic")
-            .expect("shutdown marker should be acknowledged");
-
-        let late_event = DownloadManagerEvent::Terminal(DownloadTerminalEvent::Rejected {
-            streamer_id: "streamer".to_string(),
-            streamer_name: "Streamer".to_string(),
-            session_id: "session".to_string(),
-            reason: "late event".to_string(),
-            retry_after_secs: None,
-            kind: DownloadRejectedKind::CircuitBreaker,
-        });
-        assert!(matches!(
-            late_sender.publish(late_event),
-            DownloadCoordinationReceipt::Unavailable
-        ));
     }
 }
